% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{add.phy}
\alias{add.phy}
\alias{get.phy}
\alias{remove.phy}
\title{Add, get or remove phy}
\usage{
add.phy(data, phy)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{phy}{A \code{phylo} or \code{mutiPhylo} object.}
}
\description{
Adding, extracting or removing the tree component from a \code{dispRity} object
}
\examples{
## Loading a dispRity object
data(disparity)
## Loading a tree
data(BeckLee_tree)

## Adding the tree to the dispRity object
(disparitree <- add.phy(phy = BeckLee_tree, data = disparity))

## Extracting the tree
get.phy(disparitree)

## Removing the tree from the dispRity object
remove.phy(disparitree)

}
\seealso{
\code{\link{custom.subsets}}, \code{\link{chrono.subsets}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
