% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.covar.R
\name{as.covar}
\alias{as.covar}
\title{as.covar}
\usage{
as.covar(fun, ...)
}
\arguments{
\item{fun}{a \code{function} to apply to the \code{$covar} element of \code{dispRity}.}

\item{...}{any additional arguments to pass to fun.}
}
\description{
Changes a dispRity metric to use the covar element from a dispRity object.
}
\details{
This function effectively transforms the input argument from \code{matrix} (or \code{matrix2}) to \code{matrix = matrix$VCV} and adds a evaluation after the return call to indicate that the function works on a \code{$covar} element.
Note that if the function does not have an argument called \code{matrix}, the first argument is estimated as being the one to be transformed (e.g. if the function has its first argument \code{x}, it will transform it to \code{x = x$VCV}).
}
\examples{
## Creating a dispRity
data(charadriiformes)

## Creating a dispRity object from the charadriiformes model
covar_data <- MCMCglmm.subsets(data       = charadriiformes$data,
                               posteriors = charadriiformes$posteriors)

## Get one matrix and one covar matrix
one_matrix <- get.matrix(covar_data, subsets = 1)
one_covar  <- get.covar(covar_data, subsets = 1, n = 1)[[1]][[1]]

## Measure the variances
variances(one_matrix)

## Measure the variances on the covar matrix
as.covar(variances)(one_covar)
## Is the same as:
variances(one_covar$VCV)

## Apply the measurement on a dispRity object:
## On the traitspace:
summary(dispRity(covar_data, metric = c(sum, variances))) 
## On the covariance matrices:
summary(dispRity(covar_data, metric = c(sum, as.covar(variances))))




}
\seealso{

}
\author{
Thomas Guillerme
}
