% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.target.R
\name{as.target}
\alias{as.target}
\alias{as.abundance}
\alias{as.covar}
\title{Changes the target of a metric}
\usage{
as.abundance(fun, ...)

as.covar(fun, ..., VCV, loc)
}
\arguments{
\item{fun}{a \code{function} to apply to the element of \code{dispRity}.}

\item{...}{any additional arguments to pass to fun.}

\item{VCV}{logical, whether to use the \code{$VCV} component of the elements in \code{dispRity$covar} (\code{TRUE}; default) or not (\code{FALSE}) (see details).}

\item{loc}{logical, whether to use the \code{$loc} component of the elements in \code{dispRity$covar} (\code{TRUE}) or not (\code{FALSE}; default) (see details).}
}
\description{
Changes a dispRity metric to use specific elements from the dispRity object: covar or abundance.
}
\details{
This function effectively transforms the input argument from \code{matrix} (or \code{matrix2}) to either \code{data$abundance} (for \code{as.abundance}) or \code{data$matrix$VCV} (for \code{as.covar}).

Note that if the function does not have an argument called \code{matrix}, the first argument is estimated as being the one to be transformed (e.g. if the function has its first argument \code{x}, it will transform it to \code{x = x$VCV}).

For \code{as.abundance} this function effectively transforms the input argument from \code{matrix} (or \code{matrix2}) to \code{matrix = matrix$abundance} and adds a evaluation after the return call to indicate that the function works on a \code{$abundance} element.
For \code{as.covar}, you can toggle between using the \code{$VCV} or the \code{$loc} argument in the \code{$covar} matrix by using either \code{VCV = TRUE, loc = FALSE} (to access only \code{fun(matrix = matrix$VCV, ...)}), \code{VCV = FALSE, loc = TRUE} (to access only \code{matrix = matrix(matrix$loc, nrow = 1), ...}) or \code{VCV = TRUE, loc = TRUE} (to access \code{fun(matrix = matrix$VCV, loc = matrix$loc, ...)}; provided \code{fun} has an extra \code{loc} argument).

For \code{between.groups} metrics with \code{matrix} and \code{matrix2} arguments, you can provide multiple logicals for \code{VCV} and \code{loc} to be applied repspectively to \code{matrix} and \code{matrix2}. For example \code{VCV = TRUE} will reinterpret \code{matrix} and \code{matrix2} as \code{matrix$VCV} and \code{matrix2$VCV} but \code{loc = c(TRUE, FALSE)} will only reinterpret \code{matrix} as \code{matrix$loc} (and \code{matrix2} will not be reinterpreted).
}
\examples{
## Get a dispRity object with abundance
data(BeckLee_mat50) 
abundance_data <- matrix(sample(c(0,1,2,3), 200, replace = TRUE,
                         prob = c(0.4, 0.4, 0.1, 0.1)), nrow = 50, ncol = 4)
disparabundance <- make.dispRity(data = BeckLee_mat50,
                                 abundance = abundance_data)

## Get the mean value of the traitspace ($matrix)
get.disparity(dispRity(disparabundance, metric = mean))

## Get the mean value of the abundance data ($abundance)
get.disparity(dispRity(disparabundance, metric = as.abundance(mean)))


## Get a dispRity object with covar
data(charadriiformes)

## Creating a dispRity object from the charadriiformes model
covar_data <- MCMCglmm.subsets(data       = charadriiformes$data,
                               posteriors = charadriiformes$posteriors)

## Get one matrix and one covar matrix
one_matrix <- get.matrix(covar_data, subsets = 1)
one_covar  <- get.covar(covar_data, subsets = 1, n = 1)[[1]][[1]]

## Measure the centroids
centroids(one_matrix)

## Measure the centroids on the covar matrix
as.covar(centroids)(one_covar)
## Is the same as:
centroids(one_covar$VCV)

## Apply the measurement on a dispRity object:
## On the traitspace:
summary(dispRity(covar_data, metric = c(sum, centroids))) 
## On the covariance matrices:
summary(dispRity(covar_data, metric = c(sum, as.covar(centroids))))
## The same but with additional options (centre = 100)
summary(dispRity(covar_data,
                 metric = c(sum, as.covar(centroids)),
                 centre = 100))

## Example with the VCV and loc options
## A metric that works with both VCV and loc
## (the sum of the variances minus the distance from the location)
sum.var.dist <- function(matrix, loc = rep(0, ncol(matrix))) {
    ## Get the sum of the diagonal of the matrix
    sum_diag <- sum(diag(matrix))
    ## Get the distance between 0 and the loc
    dist_loc <- dist(matrix(c(rep(0, ncol(matrix)), loc), nrow = 2, byrow = TRUE))[1]
    ## Return the sum of the diagonal minus the distance
    return(sum_diag - dist_loc)
}
## Changing the $loc on one_covar for the demonstration
one_covar$loc <- c(1, 2, 3)
## Metric on the VCV part only
as.covar(sum.var.dist, VCV = TRUE, loc = FALSE)(one_covar)
## Metric on the loc part only
as.covar(sum.var.dist, VCV = FALSE, loc = TRUE)(one_covar)
## Metric on both parts
as.covar(sum.var.dist, VCV = TRUE, loc = TRUE)(one_covar)

}
\seealso{
\code{\link{dispRity}} \code{\link{MCMCglmm.subsets}}
}
\author{
Thomas Guillerme
}
