% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cust.series.R
\name{cust.series}
\alias{cust.series}
\title{Separating ordinated data in custom series.}
\usage{
cust.series(data, series)
}
\arguments{
\item{data}{An ordinated matrix of maximal dimensions \eqn{k*(k-1)}.}

\item{series}{A \code{data.frame} with the same \eqn{k} elements as in \code{data} as rownames and a single column.}
}
\description{
Splits the ordinated data into a customized series list.
}
\details{
The customized series can typically be a factor. For a finite number of taxonomic groups, traits, etc.
}
\examples{
## Generating a dummy ordinated matrix
ordinated_matrix <- matrix(data = rnorm(90), nrow = 10, ncol = 9, dimnames = list(letters[1:10]))
## Creating a list of dummy factors (1 or 2)
factors <- as.data.frame(matrix(data = c(rep(1,5), rep(2,5)), nrow = 10, ncol = 1, dimnames = list(letters[1:10])))
## Spliting the dummy ordinated matrix
cust.series(ordinated_matrix, factors)
}
\author{
Thomas Guillerme
}

