% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity-package.R
\name{disparity}
\alias{disparity}
\title{disparity}
\format{one \code{dispRity} object.}
\description{
And example of \code{dispRity} object.
}
\details{
This matrix is based on \code{\link{BeckLee}} dataset and split into 7 continuous series (\code{\link{time.series}}).
It was bootstrapped 100 times (\code{\link{boot.matrix}}) with four rarefaction levels.
Disparity was calculated as the \code{\link[base]{median}} of the \code{\link{centroids}} (\code{\link{dispRity}}).
}

