\name{disparity.test}
\alias{dispRity.test}

\title{Some title}

\description{some description}

\usage{
disparity.test(arg)
}

\arguments{
\item{arg}{some description}

\details{If needed}

\seealso{\link{disparity} or any other functions}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}