% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity-package.R
\docType{package}
\name{dispRity-package}
\alias{dispRity-package}
\title{Measuring Disparity in R}
\description{
A modular package for measuring disparity (multidimensional space occupancy). Disparity can be calculated from any matrix defining a multidimensional space. The package provides a set of implemented metrics to measure properties of the space and allows users to provide and test their own metrics (Guillerme (2018) <doi:10.1111/2041-210X.13022>). The package also provides functions for looking at disparity in a serial way (e.g. disparity through time - Guillerme and Cooper (2018) <doi:10.1111/pala.12364>) or per groups as well as visualising the results. Finally, this package provides several statistical tests for disparity analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TGuillerme/dispRity}
}

}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\concept{disparity ordination phylogeny cladistic morphometric ecology}
