% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.BAT.R
\name{dispRity.BAT}
\alias{dispRity.BAT}
\title{Converts \code{dispRity} to \code{BAT}.}
\usage{
dispRity.BAT(data, subsets, matrix, tree)
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{dispRity} object.}

\item{subsets}{Optional, some specific subsets to extract (see \code{\link{get.subsets}}).}

\item{matrix}{Optional, some specific matrices to extract (see \code{\link{get.matrix}}).}

\item{tree}{Optional, some specific trees to extract (see \code{\link{get.tree}}).}
}
\value{
\itemize{
     \item \code{comm}
     \item \code{tree}
     \item \code{traits}
}
}
\description{
Converts a \code{dispRity} object into \code{BAT} package arguments.
}
\details{
Converts the content of a \code{dispRity} object into a list of arguments that can be used by \code{BAT} functions.
}
\examples{
## Base example:
}
