% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.covar.projections.R
\name{dispRity.covar.projections}
\alias{dispRity.covar.projections}
\title{Covar projection analyses wrapper}
\usage{
dispRity.covar.projections(
  data,
  type,
  base,
  sample,
  n,
  major.axis = 1,
  level = 0.95,
  measure = c("position", "distance", "degree"),
  verbose = FALSE
)
}
\arguments{
\item{data}{a \code{dispRity} object containing a \code{$covar} component(e.g. from \code{\link{MCMCglmm.subsets}})}

\item{type}{either \code{"groups"} for the projections between groups or \code{"elements"} for the projections of elements onto groups.}

\item{base}{optional, a specific group to project the elements or the groups onto. If left empty, the groups are projected onto each other in a pairwise manner and the elements are projected onto their respective groups.}

\item{sample}{optional, one or more specific posterior sample IDs (is ignored if n is used) or a function to summarise all axes.}

\item{n}{optional, a random number of covariance matrices to sample (if left empty, all are used).}

\item{major.axis}{which major axis to use (default is \code{1}; see \code{\link{axis.covar}} for more details).}

\item{level}{the confidence interval to estimate the major axis (default is \code{0.95}; see \code{\link{axis.covar}} for more details)).}

\item{verbose}{logical, whether to be verbose (\code{TRUE}) or not (\code{FALSE}, default).}

\item{output}{which values to output from the projection. By default, the three values \code{c("position", "distance", "degree")} are used to respectively output the projection, rejection and angle values (see \code{\link{projections}} for more details).}
}
\value{
A \code{list} of \code{dispRity} objects corresponding to each projection value from \code{output}.
To visualise/summarise each element you can either select them by name (e.g. \code{output$position}) or by ID (e.g. \code{output[[1]]}).
}
\description{
Wrapper function for a covar projection analyses on dispRity objects
}
\details{
Effectively, the wrapper runs either of the following function (simplified here):
\itemize{
     \item if \code{type = "groups"}: \code{dispRity(data, metric = as.covar(projections.between), between.groups = TRUE, )} for the projections group in \code{data} onto each other. 
     \item if \code{type = "elements"}: \code{dispRity(data, metric = as.covar(projections), ...)} for the projections of each element in \code{data} onto their main axis.
}

If \code{base} is specified:
\itemize{
     \item \code{type = "groups"} will run pairs elements each subset and \code{base} (instead of the full pairwise analyses).
     \item \code{type = "elements"} will run the projection of each subset onto the major axis from \code{base} rather than its own.
}
}
\examples{

## TODO:
example
1 = 2





}
\seealso{
\code{\link{projections}} \code{\link{projections.elements}} \code{\link{axis.covar}} \code{\link{dispRity}} \code{\link{MCMCglmm.subsets}}
}
\author{
Thomas Guillerme
}
