% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.fast.R
\name{dispRity.fast}
\alias{dispRity.fast}
\title{Fast dispRity}
\usage{
dispRity.fast(group, space, metric)
}
\arguments{
\item{group}{a logical vector for grouping}

\item{space}{a matrix}

\item{metric}{a metric dispRity style (up to two levels)}
}
\description{
Fast disparity calculations. THIS FUNCTION IS LESS SAFE TO USE THAN \code{\link{dispRity}} (see details).
}
\details{
\emph{IN DOUBT, USE \code{\link{dispRity}} INSTEAD OF THIS FUNCTION.}
This function should only be used in very specific cases requiring advanced optimisation or embedded customised functions.
This function is simply applying \code{metric(space[group,])} for each group and returns a list of results.
It does not check the validity of the data, metric and groups.
It does not handle specific data (e.g. multiple matrices), specific metrics (e.g. no optional arguments), does not produce meaningful error messages and does not intake nor returns a dispRity object.
}
\examples{
## A random space
space <- matrix(rnorm(25), 5, 5)

## A metric
metric <- c(sum, variances)

## A group of four observations
group <- c(TRUE, TRUE, TRUE, TRUE, FALSE)

## The disparity
dispRity.fast(group, space, metric)


}
\author{
Thomas Guillerme
}
