% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.wrapper.R
\name{dispRity.per.group}
\alias{dispRity.per.group}
\title{Disparity per groups.}
\usage{
dispRity.per.group(data, group, metric = c(median, centroids), ...)
}
\arguments{
\item{data}{An ordinated \code{matrix}.}

\item{group}{A \code{list} of row numbers for each group.}

\item{metric}{A vector containing one to three functions (default = \code{c(median, centroids)}) (see \code{\link{dispRity}} for details).}

\item{...}{Optional arguments to be passed to \code{\link{custom.subsamples}}, \code{\link{boot.matrix}} and \code{\link{dispRity}}.}
}
\value{
A \code{dispRity} object that can be passed to \code{summary} or \code{plot}.
}
\description{
Performs a disparity analysis between groups.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Run a simple disparity through time analysis (with 3 time bins)
result <- dispRity.per.group(BeckLee_mat50, list(crown = c(16, 19:41, 45:50),
                             stem = c(1:15, 17:18, 42:44)))
summary(result) ; plot(result)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{custom.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity.metric}}, \code{\link{summary.dispRity}}, \code{\link{plot.dispRity}}.
}

