\name{disparity}
\alias{dispRity}

\title{Calculates disparity from an ordinated matrix}

\description{some description}

\usage{
disparity(data, metric, bootstrap, rarefaction, diversity, bs.method)
}

\arguments{
\item{data}{some description}
\item{metric}{some description}
\item{bootstrap}{some description}
\item{rarefaction}{some description}
\item{diversity}{some description}
\item{bs.method}{some description}

\details{If needed}

\seealso{\link{disparity} or any other functions}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}