% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispRity.test.R
\name{dispRity.test}
\alias{bhatt.coeff}
\alias{dispRity.test}
\title{Disparity tests}
\usage{
bhatt.coeff(x, y, bw = bw.nrd0, ...)
}
\arguments{
\item{x,y}{two distributions.}

\item{bw}{the bandwith size, either a \code{numeric} or a \code{function} (see \code{\link[stats]{bw.nrd0}}).}

\item{...}{optional arguments to be passed to \code{bw}.}
}
\description{
Different implemented tests for comparing series.
}
\details{
These are inbuilt statistical tests for comparing disparity series.
\itemize{
  \item \code{bhatt.coeff}: calculates the Bhattacharyya Coefficient (probability of overlap) between two distributions.
}
}
\examples{
##
}
\author{
Thomas Guillerme
}
\references{
Bhattacharyya A. 1943. On a measure of divergence between two statistical populations defined by their probability distributions. Bull. Calcutta Math. Soc., \bold{35}, pp. 99–-109
}
\seealso{
\code{\link{test.dispRity}}.
}

