% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{extract.dispRity}
\alias{extract.dispRity}
\title{Extracting disparity values.}
\usage{
extract.dispRity(data, observed = TRUE, rarefaction, series)
}
\arguments{
\item{data}{A \code{dispRity} object containing disparity results.}

\item{observed}{A \code{logical} value indicating whether to output the observed (\code{TRUE} (default)) or the bootstrapped values (\code{FALSE}).}

\item{rarefaction}{Optional, a single \code{numeric} value corresponding to the rarefaction level (as the number of elements; if missing, the non-rarefied values are output).}

\item{series}{Optional, a \code{numeric} or \code{character} for which series to get (if missing, the value for all series are given).}
}
\description{
Extracts the disparity from a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat99) ; data(BeckLee_tree)

## Calculating some disparity
series <- time.series(BeckLee_mat99, BeckLee_tree, method = "continuous",
     time = c(100, 80, 60), model = "gradual")
bootstraps_dat <- boot.matrix(series, bootstraps = 10, rarefaction = TRUE)
disparity_data <- dispRity(bootstraps_dat, metric = mean)

## Extracting the observed disparity
extract.dispRity(disparity_data)

## Extracting the bootstrapped disparity
boot_disp <- extract.dispRity(disparity_data, observed = FALSE)
str(boot_disp)
## Or with the minimum rarefaction
boot_disp_rare <- extract.dispRity(disparity_data, observed = FALSE,
     rarefaction = "min")
str(boot_disp_rare)
## Or a certain level of rarefaction
boot_disp_rare <- extract.dispRity(disparity_data, observed = FALSE,
     rarefaction = 5)
str(boot_disp_rare)

## Extracting bootstrapped distributions of disparity
disparity_distr <- dispRity(bootstraps_dat, metric = variances)
boot_disp <- extract.dispRity(disparity_distr, observed = FALSE,
     concatenate = FALSE, keep.structure = TRUE)
str(boot_disp) # A list of three lists of 10 elements

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{get.dispRity}}.
}

