% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.dispRity.R
\name{fetch.dispRity}
\alias{fetch.dispRity}
\alias{fetch.elements}
\alias{fetch.matrix}
\alias{fetch.series}
\title{Fetching elements of a \code{dispRity} object.}
\usage{
fetch.matrix(dispRity, series, rarefaction, bootstrap)
fetch.elements(dispRity, series)
fetch.series(dispRity)
}
\arguments{
\item{dispRity}{A \code{dispRity} object.}

\item{series}{A \code{numeric} value to select a series (\code{0} is no series; default).}

\item{rarefaction}{A \code{numeric} value to select the rarefaction level (\code{0} is no rarefaction; default).}

\item{bootstrap}{A \code{numeric} value to select a specific bootstrap draw (\code{0} is no bootstrap; default).}
}
\description{
Fetching the matrix, elements or series of a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping and rarefying the data
dispRity_object <- boot.matrix(customised_series, bootstraps = 100,
     rarefaction = c(15, 10))

## To get the original matrix
fetch.matrix(dispRity_object, 0,0,0)
# or simply fetch.matrix(dispRity_object)
## To get the matrix of the second series, first rarefaction, 58th bootstrap
fetch.matrix(dispRity_object, 2,1,58)
## To get the elements in the first series
fetch.elements(dispRity_object, 1)
## To get the series
fetch.series(dispRity_object)

}
\author{
Thomas Guillerme
}

