% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{fill.dispRity}
\alias{fill.dispRity}
\title{Fills a \code{dispRity} object.}
\usage{
fill.dispRity(data)
}
\arguments{
\item{data}{A \code{dispRity} object.}
}
\description{
Fills a \code{dispRity} object using the data from its matrix
}
\examples{
## An empty dispRity object (with a matrix)
empty <- make.dispRity(data = matrix(rnorm(12), ncol = 3))

## A dispRity object with a matrix of 4*3
fill.dispRity(empty)

}
\author{
Thomas Guillerme
}
