% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.ordination.R
\name{geomorph.ordination}
\alias{geomorph.ordination}
\title{Imports data from geomorph}
\usage{
geomorph.ordination(data, k, ...)
}
\arguments{
\item{data}{An array (p x k x n) containing landmark coordinates for a set of specimens.}

\item{k}{The number of dimensions in the ordination. If left empty, the number of dimensions is set to number of rows - 1.}

\item{...}{Any optional arguments to be passed to \code{\link[geomorph]{gpagen}}.}
}
\description{
Takes geomorph landmark data and computes both a procrustes analysis and an ordination
}
\seealso{
\code{\link[geomorph]{gpagen}}, \code{\link[stats]{cmdscale}}, \code{\link{custom.subsamples}}, \code{\link{time.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}

