% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dispRity.R
\name{get.dispRity}
\alias{get.dispRity}
\title{Exatracting subsamples from a dispRity object.}
\usage{
get.dispRity(data, what, keep.elements = FALSE)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{what}{A list of series names or series numbers to be extracted.}

\item{keep.elements}{\code{logical}, whether to keep the elements from the original matrix (default = \code{FALSE}).}
}
\value{
This function outputs a \code{dispRity} object containing:
\item{data}{A \code{list} of the observed and boostraped matrices.}
\item{disparity}{A \code{list} of disparity values.}
\item{elements}{A \code{vector} containing all the names of the elements from the original matrix.}
\item{series}{A \code{vector} containing the name of the series (is \code{"1"} if the input was a single \code{matrix}).}
\item{call}{A \code{vector} containing the arguments used for the bootstraping.}
\code{dispRity} objects can be summarised using \code{print} (S3).
Use \link{summary.dispRity} to summarise the \code{dispRity} object.
}
\description{
Extrancting subsamples series and data from a dispRity object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat99) ; data(BeckLee_tree) 

## Series subsamples
series_full <- time.series(BeckLee_mat99, BeckLee_tree, method="continuous", time=5, model="acctran")
series_full # 5 series for 99 elements
get.dispRity(series_full, what=1) # 1 series for 3 elements

## Boostrapped data subsamples
bootstrapped_data <- boot.matrix(series_full, bootstraps=10, rarefaction=c(3,5))
bootstrapped_data # 5 series for 99 elements
get.dispRity(bootstrapped_data, what="66.75552") # 1 series for 23 elements

## Disparity data subsamples
disparity_data <- dispRity(bootstrapped_data, c(sum, ranges))
disparity_data # 5 series for 99 elements
get.dispRity(disparity_data, what=c(1,5)) # 2 series for 13 elements

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{cust.series}}, \code{\link{time.series}}, \code{\link{boot.matrix}} and \code{\link{dispRity}}.
}

