% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinction.reference.R,
%   R/extinction.reference_fun.R
\name{get.extinction.reference}
\alias{get.extinction.reference}
\alias{get.extinction.reference}
\title{Getting the reference and comparison slices}
\usage{
get.extinction.reference(data, extinction, lag = 1, names = FALSE)

get.extinction.reference(data, extinction, lag = 1, names = FALSE)
}
\arguments{
\item{data}{a dispRity object}

\item{extinction}{the extinction time}

\item{lag}{the lag effect (i.e. how many bins/slices after the extinction to consider)}

\item{names}{whether to display the bins names (\code{TRUE}) or not (\code{FALSE} - default)}

\item{data}{a dispRity object}

\item{extinction}{the extinction time}

\item{lag}{the lag effect (i.e. how many bins/slices after the extinction to consider)}

\item{names}{whether to display the bins names (\code{TRUE}) or not (\code{FALSE} - default)}
}
\description{
Getting the reference (pre-extinction) and the comparison (post-extinction) slices/bins

Getting the reference (pre-extinction) and the comparison (post-extinction) slices/bins
}
\examples{


}
\seealso{



}
\author{
Thomas Guillerme

Thomas Guillerme
}
