% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{get.series.dispRity}
\alias{get.dispRity}
\alias{get.series.dispRity}
\title{Extracts series from a dispRity object.}
\usage{
get.series.dispRity(data, series)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{series}{A list of series names or series numbers to be extracted.}
}
\value{
This function outputs a \code{dispRity} object.
}
\description{
Extracting some series and data from a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat99) ; data(BeckLee_tree) 

## Series sub-samples
series_full <- time.series(BeckLee_mat99, BeckLee_tree,
     method = "continuous",time = 5, model = "acctran")
series_full # 5 series for 99 elements

get.series.dispRity(series_full, 1) # 1 series for 3 elements

## Bootstrapped data sub-samples
bootstrapped_data <- boot.matrix(series_full, bootstraps = 10,
     rarefaction = c(3, 5))
bootstrapped_data # 5 series for 99 elements
get.series.dispRity(bootstrapped_data, "66.75552") # 1 series for 23 elements

## Disparity data sub-samples
disparity_data <- dispRity(bootstrapped_data, variances)
disparity_data # 5 series for 99 elements
get.series.dispRity(disparity_data, c(1,5)) # 2 series for 13 elements

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{extract.dispRity}}.
}

