% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{get.subsamples}
\alias{get.dispRity,}
\alias{get.subsamples}
\alias{get.subsamples.dispRity}
\title{Extracts subsamples from a dispRity object.}
\usage{
get.subsamples(data, subsamples)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{subsamples}{A list of subsamples names or subsamples numbers to be extracted.}
}
\value{
This function outputs a \code{dispRity} object.
}
\description{
Extracting some subsamples and data from a \code{dispRity} object.
}
\examples{
## Load the disparity data based on Beck & Lee 2014
data(disparity)

## Get one subsamples
get.subsamples(disparity, "60")

## Get two subsamples
get.subsamples(disparity, c(1,5))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{extract.dispRity}}.
}

