% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.dispRity.models.R
\name{make.dispRity.models}
\alias{make.dispRity.models}
\title{Making models for changes in disparity}
\usage{
make.dispRity.models(model)
}
\arguments{
\item{model}{A \code{function} to be passed as a model to \code{\link{model.test}}}

\item{...}{Any additional options to be passed to this model.}
}
\description{
Testing whether a user defined model for testing dispRity mode of evolution works
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
