% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{matrix.dispRity}
\alias{fetch.matrix}
\alias{matrix.dispRity}
\title{Fetching a matrix from a \code{dispRity} object.}
\usage{
matrix.dispRity(data, series, rarefaction, bootstrap)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{series}{A \code{numeric} value to select a series (\code{0} is no series; default).}

\item{rarefaction}{A \code{numeric} value to select the rarefaction level (\code{0} is no rarefaction; default).}

\item{bootstrap}{A \code{numeric} value to select a specific bootstrap draw (\code{0} is no bootstrap; default).}
}
\description{
Fetching a specific matrix from a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping and rarefying the data
dispRity_data <- boot.matrix(customised_series, bootstraps = 100,
     rarefaction = c(15, 10))

## To get the original matrix
matrix.dispRity(dispRity_data)

## To get the un-bootstrapped matrix from the second series
matrix.dispRity(dispRity_data, series = 2)

## To get the 52nd bootstrap draw of the second rarefaction level (15) of the
## same series
matrix.dispRity(dispRity_data, series = 2, rarefaction = 2, bootstrap = 52)

}
\author{
Thomas Guillerme
}

