% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity-package.R
\name{McClean_data}
\alias{McClean_data}
\title{McClean dataset}
\format{one matrix and two vectors.}
\description{
Example datasets from McClean (unpublished).
}
\details{
\itemize{
  \item \code{ordination} The ordinated matrix based on the 40 sites and 20 dimensions
  \item \code{treatment} A vector of \code{character} treatment types (\code{"a"} or \code{"b"})
  \item \code{depth} A vector of \code{numeric} depth types (\code{1} or \code{2})
}
}
\seealso{
BeckLee_data disparity
}

