% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCglmm.subsets.R
\name{MCMCglmm.subsets}
\alias{MCMCglmm.subsets}
\title{MCMCglmm.subsets}
\usage{
MCMCglmm.subsets(data, posteriors, group, tree, rename.groups)
}
\arguments{
\item{data}{The data used for the MCMCglmm model}

\item{posteriors}{A MCMCglmm object, the posteriors of the model}

\item{group}{Optional, which group to include from the posteriors (if left empty the residuals and random terms are used)}

\item{tree}{Optional, the tree(s) used in the MCMCglmm analyses}

\item{rename.groups}{optional, a vector of group names for renaming them.}
}
\description{
Creating a dispRity object from a MCMCglmm posterior output
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
