% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCglmm.utilities.R
\name{MCMCglmm.utilities}
\alias{MCMCglmm.utilities}
\alias{MCMCglmm.levels}
\alias{MCMCglmm.traits}
\alias{MCMCglmm.sample}
\alias{MCMCglmm.covars}
\title{MCMCglmm object utility functions}
\usage{
MCMCglmm.traits(MCMCglmm)

MCMCglmm.levels(MCMCglmm, convert)

MCMCglmm.sample(MCMCglmm, n)

MCMCglmm.covars(MCMCglmm, n, sample)
}
\arguments{
\item{MCMCglmm}{A \code{MCMCglmm} object.}

\item{convert}{Logical, whether to return the raw term names names as expressed in the model column names (\code{FALSE}) or to convert it to something more reader friendly (\code{TRUE}; default).}

\item{n}{Optional, a number of random samples to extract.}

\item{sample}{Optional, the specific samples to extract (is ignored if \code{n} is present).}
}
\description{
Different utility functions to extract aspects of a \code{MCMCglmm} object.
}
\details{
\itemize{
     \item \code{MCMCglmm.levels} returns the different random and residual terms levels of a \code{MCMCglmm} object. This function uses the default option \code{convert = TRUE} to convert the names into something more readable. Toggle to \code{convert = FALSE} for the raw names.
     \item \code{MCMCglmm.traits} returns the column names of the different traits of a \code{MCMCglmm} formula object.
     \item \code{MCMCglmm.sample} returns a vector of sample IDs present in the \code{MCMCglmm} object. If \code{n} is missing, all the samples IDs are returned. Else, a random series of sample IDs are returned (with replacement if n greater than the number of available samples).
     \item \code{MCMCglmm.covars} returns a list of covariance matrices and intercepts from a \code{MCMCglmm} object (respectively from \code{MCMCglmm$VCV} and \code{MCMCglmm$Sol}). By default, all the covariance matrices and intercepts are returned but you can use either of the arguments \code{sample} to return specific samples (e.g. \code{MCMCglmm.covars(data, sample = c(1, 42))} for returning the first and 42nd samples) or \code{n} to return a specific number of random samples (e.g. \code{MCMCglmm.covars(data, n = 42)} for returning 42 random samples).
}
}
\examples{
## Loading the charadriiformes model
data(charadriiformes)
model <- charadriiformes$posteriors
class(model) # is MCMCglmm

## Get the list of levels from the model
MCMCglmm.levels(model)
## The raw levels names (as they appear in the MCMCglmm object)
MCMCglmm.levels(model, convert = FALSE)

## Get the traits names from the model
MCMCglmm.traits(model)

## Get all the available samples in the model
length(MCMCglmm.sample(model))
## Get 5 random sample IDs from the model
MCMCglmm.sample(model, n = 5)

## Get one specific samples from the model
MCMCglmm.covars(model, sample = 42)
## Get two random samples from the model
MCMCglmm.covars(model, n = 2)
}
\seealso{
\code{\link{MCMCglmm.subsets}}
}
\author{
Thomas Guillerme
}
