% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.time.series.R
\name{merge.time.series}
\alias{merge.time.series}
\title{Combines or cleans time series.}
\usage{
\method{merge}{time.series}(data1, data2, after = TRUE)
}
\arguments{
\item{data1}{An ordinated matrix of maximal dimensions \eqn{k*(k-1)}, or a \code{dispRity} object (see details).}

\item{data2}{An optional matrix of maximal dimensions \eqn{k*(k-1)}, or a \code{dispRity} object (see details). If missing \code{data1} will be cleaned.}

\item{after}{A \code{logical} value indicating whether to be merge before or after the first series (\code{default = TRUE}).}
}
\value{
This function outputs a \code{dispRity} object containing:
\item{data}{A \code{list} of the split ordinated data (each element is a \code{matrix}).}
\item{elements}{A \code{vector} containing all the rownames from the input matrix.}
\item{series}{A \code{vector} containing the name of the series.}
\code{dispRity} objects can be summarised using \code{print} (S3).
}
\description{
Combines two time series sequentially or cleans a single time series by merging any series with less than 3 taxa to the previous or next series.
}
\details{
The \code{dispRity} object given to the \code{data} argument can be: a list of matrices (typically output from the functions \code{\link{time.series}} or \code{\link{cust.series}}), a bootstrapped matrix output from \code{\link{boot.matrix}} or a list of disparity measurements calculated from this \code{dispRity} function.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{cust.series}}, \code{\link{time.series}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}

