% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.R
\name{model.test.BM}
\alias{model.test.BM}
\title{BM Model Testing}
\usage{
model.test.BM(data, ...)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{...}{the model parameters}
}
\description{
Applies a Brownian Motion model to disparity data
}
\examples{

}
\seealso{
\link{\code{model.test}}
}
\author{
Mark Puttick and Thomas Guillerme
}
