% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.R
\name{model.test.EB}
\alias{model.test.EB}
\title{EB Model Testing}
\usage{
model.test.EB(data, ...)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{...}{the model parameters}
}
\description{
Applies an Early Burst model to disparity data
}
\examples{

}
\seealso{
\link{\code{model.test}}
}
\author{
Mark Puttick and Thomas Guillerme
}
