% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.R
\name{model.test.Shift}
\alias{model.test.Shift}
\title{Shift Model Testing}
\usage{
model.test.Shift(data, models, parameters, ...)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{models}{the name of the models to shift (\code{"EB"}, \code{"BM"}, \code{"OU"} or \code{"Trend"}).}

\item{parameters}{a \code{list} of model parameters to be passed to each model.}

\item{shifts}{where to apply the shifts.}
}
\description{
Applies a Shift model to disparity data
}
\examples{

}
\seealso{
\link{\code{model.test}}
}
\author{
Mark Puttick and Thomas Guillerme
}
