% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{n.subsets}
\alias{n.subsets}
\title{Number of subsets.}
\usage{
n.subsets(data)
}
\arguments{
\item{data}{A \code{dispRity} object.}
}
\description{
Getting the number of subsets from a dispRity object.
}
\examples{
## Loading a dispRity object
data(disparity)

## What are the number of elements per subsets?
n.subsets(disparity)

}
\seealso{
\code{\link{custom.subsets}}, \code{\link{chrono.subsets}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
