% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.test.R
\name{null.test}
\alias{null.test}
\title{Testing a null hypothesis on multidimensional data.}
\usage{
null.test(data, replicates, null.distribution, null.arguments = NULL, alter = "two-sided", ...)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{replicates}{the number of replicates for the test.}

\item{null.distrib}{one or more distribution functions to generate the null model to be passed to \code{\link{space.maker}}.}

\item{null.args}{any additional arguments to be passed to \code{\link{space.maker}} (\code{default = NULL}).}

\item{alter}{the type of alternative hypothesis (H1) as used in \code{\link[ade4]{randtest}} (\code{default = "two-sided"}).}

\item{scale}{whether to scale the simulated and the observed data.}

\item{...}{optional arguments to be passed to \code{\link[ade4]{as.randtest}}.}
}
\description{
Testing the difference between the observed disparity and disparity under a null model.
}
\details{

}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)
## Calculating the disparity as the ellipsoid volume
obs_disparity <- dispRity(BeckLee_mat50, metric = ellipse.volume)
## Testing 
results <- null.test(obs_disparity, replicates = 100, null.distrib = rnorm)
results
plot(results)

}
\author{
Thomas Guillerme
}
\references{
Díaz, S., Kattge, J., Cornelissen, J.H., Wright, I.J., Lavorel, S., Dray, S., Reu, B., Kleyer, M., Wirth, C., Prentice, I.C. and Garnier, E., \bold{2016}. The global spectrum of plant form and function. Nature, 529(7585), pp.167-171.
}
\seealso{
\code{\link{space.maker}}, \code{\link{test.dispRity}}, \code{\link{sequential.test}}, \code{\link{sequential.test}}.
}

