% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.plot.R
\name{pair.plot}
\alias{pair.plot}
\title{Plots pairwise comparisons with}
\usage{
pair.plot(data, what, col = c("black", "white"), legend = FALSE, binary,
  diag, add, ...)
}
\arguments{
\item{data}{A \code{data.frame} object with comparisons pair names as row names.}

\item{what}{A \code{numeric} or \code{character} value designating which column to plot.}

\item{col}{The two extremes of a color gradient (default = \code{c("black", "white")}).}

\item{legend}{Logical, whether to plot the legend or not.}

\item{binary}{Optional, if the results must be binary, a \code{numeric} value for the threshold of acceptance (values greater will be 1, lower will be 0).}

\item{diag}{Optional, can be \code{"max"} or \code{"min"} or a single \code{numeric} value.}

\item{add}{Optional, whether to add significance tokens can be \code{numeric} for a point type to print (\code{pch}) or \code{"character"} to print (e.g. \code{"*"}).}

\item{...}{Any other options to be passed to \code{\link[graphics]{plot}}.}
}
\description{
Plots pairwise comparisons from a data frame (typically output from \code{link{test.dispRity}}).
}
\examples{
## A small data.frame
data <- data.frame(matrix(data = runif(40), ncol = 2))

## Plotting the first column as a pairwise comparisons
pair.plot(data, what = 1, col = c("orange", "blue"), legend = TRUE, diag = 1)

## Adding some tokens for each value below 0.2 in the second column
pair.plot(data, what = 2, binary = 0.2, add = "*", cex = 2)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{test.dispRity}}.
}

