% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.space.R
\name{reduce.space}
\alias{reduce.space}
\title{Reduce space}
\usage{
reduce.space(
  space,
  type,
  remove,
  parameters,
  tuning,
  verbose = FALSE,
  return.optim = FALSE
)
}
\arguments{
\item{space}{the trait space}

\item{type}{how to reduce the space (either \code{"random"}, \code{"limit"}, \code{"displacement"} or \code{"density"})}

\item{remove}{the proportion of elements to be removed (in probability)}

\item{parameters}{the parameter(s) for removal selection (see details). If left empty, the \code{parameters} is estimated to reach the amount set by \code{remove}.}

\item{tuning}{Optinal parameters for tuning the parameter estimations (if remove is required and parameters is missing) a list of three parameters: "max" for the maximum of operations, "tol" for the tuning (e.g. 0.1 close), "inc.steps" for the initial increment value during optimisation (default = 2 - the bigger the value, the slower the increment).}

\item{verbose}{wether to be verbose or not}

\item{return.optim}{logical, whether to also return the optimal value.}
}
\description{
Remove elements from a multidimensional space
}
\details{
- \code{limit.removal parameters}: a list of \code{parameters$centre}, the centre from which to count the radius (if missing, is set to \code{0}); and \code{parameters$radius}, the radius for removal.

- \code{displacement.removal parameters}: a list of \code{parameters$value}, value the threshold value from which to remove elements.

- \code{density.removal parameters}: a list of \code{parameters$what} "close" (default) for close neighbours or "distant" for distant ones; \code{parameters$diameter} the diameter for considering closeness or distance; \code{parameters$output} either "singles" or "pairs" to return the pairs of neighbours or one of them only (the first).

See Guillerme et al. 2020 and https://github.com/TGuillerme/moms for details.
}
\examples{
set.seed(1)
## Creating a two dimensional space
space <- dispRity::space.maker(100, 2, distribution = stats::rnorm)

## Generating the four types of reductions
random <- reduce.space(space, "random", remove = 0.5)
limit <- reduce.space(space, "limit", remove = 0.5)
displacement <- reduce.space(space, "displacement", remove = 0.5)
density <- reduce.space(space, "density", remove = 0.5)

## Plotting the four different results
par(mfrow = c(2,2))
plot(space, pch = 19, col = c("grey", "black")[as.factor(random)],
     main = "Random removal") 
plot(space, pch = 19, col = c("grey", "black")[as.factor(limit)],
     main = "Limit removal")
plot(space, pch = 19, col = c("grey", "black")[as.factor(displacement)],
     main = "Displacement removal")
plot(space, pch = 19, col = c("grey", "black")[as.factor(density)],
     main = "Density removal")


}
\references{
Guillerme T, Puttick MN, Marcy AE, Weisbecker V. \bold{2020} Shifting spaces: Which disparity or dissimilarity measurement best summarize occupancy in multidimensional spaces?. Ecol Evol. 2020;00:1-16. (doi:10.1002/ece3.6452)
}
\author{
Thomas Guillerme
}
