% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.zero.brlen.R
\name{remove.zero.brlen}
\alias{remove.zero.brlen}
\title{Remove zero branch length}
\usage{
remove.zero.brlen(tree, slide, verbose = FALSE)
}
\arguments{
\item{tree}{A \code{"phylo"} object with edge lengths}

\item{slide}{An optional sliding \code{numeric} values. If left empty, 1\% of the shortest branch length is used.}

\item{verbose}{A \code{logical} value indicating whether to be verbose or not.}
}
\value{
A \code{"phylo"} object with a postorder edge table and no zero branch lengths.
}
\description{
Remove zero branch lengths on trees by sliding nodes randomly in a postorder traversal based on \code{\link{slide.nodes}}.
}
\details{
The sliding value will be used to slide the nodes up and down to remove zero branch lengths by minimising the amount of branch changes.
The algorithm slides the nodes up and down (when possible) on each node in a recursive way while there is still zero branch lengths.
If two recursions produce the same series of zero branches (e.g. by sliding node A towards node B equally so that the distance A:B becomes 0), the sliding value is divided by two until the next slide.
}
\examples{

}
\seealso{
\code{\link{slide.nodes}}
}
\author{
Thomas Guillerme
}
