% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{size.subsamples}
\alias{size.subsamples}
\title{Size subsamples.}
\usage{
size.subsamples(data)
}
\arguments{
\item{data}{A \code{dispRity} object.}
}
\description{
Getting the size (number of elements) from each subsamples of a dispRity object.
}
\examples{
## Loading a dispRity object
data(disparity)

## What are the number of elements per subsamples?
size.subsamples(disparity)

}
\seealso{
\code{\link{custom.subsamples}}, \code{\link{time.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
