% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{size.subsets}
\alias{size.subsets}
\title{Size subsets.}
\usage{
size.subsets(data)
}
\arguments{
\item{data}{A \code{dispRity} object.}
}
\description{
Getting the size (number of elements) from each subsets of a dispRity object.
}
\examples{
## Loading a dispRity object
data(disparity)

## What are the number of elements per subsets?
size.subsets(disparity)

}
\seealso{
\code{\link{custom.subsets}}, \code{\link{chrono.subsets}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
