% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.maker.R
\name{space.maker}
\alias{space.maker}
\title{Creating some multidimensional spaces}
\usage{
space.maker(elements, dimensions, distribution, ...)
}
\arguments{
\item{elements}{An \code{numeric} value.}

\item{dimensions}{An \code{numeric} value smaller than \code{elements}.}

\item{distribution}{One or more \code{function} to determine the distribution of the \code{elements} along each \code{dimensions}. The function must have a single input: \code{elements}.}

\item{...}{Optional arguments to be passed to function.}
}
\description{
Creates a multidimensional space with a given number of elements and dimensions
}
\details{

}
\examples{
## A square space
plot(space.maker(5000, 2, runif), pch=20)

## A circular space
plot(space.maker(5000, 2, rnorm), pch=20)

## A 2D cilindrical space
plot(space.maker(5000, 2, c(rnorm, runif)), pch=20)

## Not run:
## A cube space
require(scatterplot3d)
scatterplot3d(space.maker(5000, 3, runif), pch=20)

## A sphere space
scatterplot3d(space.maker(5000, 3, rnorm), pch=20)

## A 3D cilindrical space
scatterplot3d(space.maker(5000, 3, c(rnorm, rnorm, runif)), pch=20)

## End(Not run)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{make.metric}}.
}

