% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{t.dispRity}
\alias{t.dispRity}
\title{Matrix transpose for \code{dispRity}.}
\usage{
\method{t}{dispRity}(x)
}
\arguments{
\item{x}{A \code{dispRity} object.}
}
\description{
Transpose the matrices in a \code{dispRity} object.
}
\details{
Transposes the matrix or matrices in the \code{dispRity} object as well as all other optional linked data (e.g. dimensions, abundance data, etc) if present.
}
\examples{
## Load a disparity example dataset
data(disparity)

## Transpose the internal data (subsets and disparity are lost)
t(disparity)

}
\seealso{
\code{\link{dispRity}}
}
\author{
Thomas Guillerme
}
