\name{test.diff}
\alias{test.difference}

\title{Testing difference between series.}

\description{Testing the statistical difference between disparity series.}

\usage{
test.diff(dispRity, test, ...)
}

\arguments{
\item{dispRity}{a dispRity object}

\details{If needed}

\seealso{\link{dispRity} or any other functions}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}