\name{time.series}
\alias{time.series}

\title{Separating ordinated data in time series.}

\description{Splits the ordinated data into a time series list.}

\usage{
time.series(data, tree, method, time, model, inc.nodes, FADLAD)
}

\arguments{
\item{data}{some description}
\item{tree}{some description}
\item{method}{some description}
\item{time}{some description}
\item{model}{some description}
\item{inc.nodes}{some description}
\item{FALAD}{optional, some description}

\details{If needed}

\seealso{\link{cust.series} for custom time series.}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}