% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{add.tree}
\alias{add.tree}
\alias{get.tree}
\alias{remove.tree}
\title{Add, get or remove tree}
\usage{
add.tree(data, tree)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{tree}{A \code{phylo} or \code{mutiPhylo} object.}
}
\description{
Adding, extracting or removing the tree component from a \code{dispRity} object
}
\examples{
## Loading a dispRity object
data(disparity)
## Loading a tree
data(BeckLee_tree)

## Adding the tree to the dispRity object
(disparitree <- add.tree(tree = BeckLee_tree, data = disparity))

## Extracting the tree
get.tree(disparitree)

## Removing the tree from the dispRity object
remove.tree(disparitree)

}
\seealso{
\code{\link{custom.subsets}}, \code{\link{chrono.subsets}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
