% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.metric.R
\name{BAT.metric}
\alias{BAT.metric}
\title{Using metrics from the BAT package.}
\usage{
BAT.metric(matrix, BAT.fun, ...)
}
\arguments{
\item{matrix}{A data matrix.}

\item{BAT.fun}{The name of the metric or directly it's function.}

\item{...}{Any optional arguments to be passed to \code{BAT.metric}}
}
\description{
An interface to use metrics from the \code{BAT} package in \code{dispRity}
}
\details{
This function is an interface between the \code{dispRity} and B\code{BAT}AT package allowing to use a \code{BAT} function in the \code{dispRity} pipeline.
By default the function uses the \code{dispRity.metric} format:

\code{metric_value <- BAT.metric(matrix, BAT.fun = BAT::alpha, ...)}

or

\code{metric_value <- BAT.metric(matrix, BAT.fun = "alpha", ...)} 

With any optional argument being handled normally.
However, most commonly, the function can be used in the context of the \code{dispRity} pipeline as follows:

\code{my_disparity <- dispRity(my_data, metric = BAT.metric, BAT.fun = BAT::alpha, ...)}

or

\code{metric_value <- dispRity(my_data, metric = BAT.metric, BAT.fun = "alpha", ...)} 

The default argument \code{tree} is handled differently depending on the context:
\itemize{
     \item if...
     \item if...
     \item if...
}
}
\examples{
## Base example:
## Generating a matrix
dummy_matrix <- matrix(rnorm(90), 9, 10)

## Applying a BAT metric to it
alpha_diversity <- BAT.metric(dummy_matrix, BAT.fun = "alpha")

## dispRity example:
## Load ecological data
data(demo_data)
eco_data <- demo_data$jones
## Subseted data based on two groups
eco_data
## Apply the alpha diversity on these subsets
alpha_diversity <- dispRity(eco_data, metric = BAT.metric, BAT.fun = "alpha")
summary(alpha_diversity)
}
\seealso{
\code{\link{dispRity}}, \code{\link{custom.subsets}}
}
\author{
Thomas Guillerme
}
