% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.metric.R
\name{BAT.metric}
\alias{BAT.metric}
\title{Using metrics from the BAT package.}
\usage{
BAT.metric(matrix, BAT.fun, BAT.args = NULL, return.raw = FALSE)
}
\arguments{
\item{matrix}{A data matrix or a \code{BAT} structure data list (containing \code{"comm"}, \code{"tree"}, \code{"traits"}).}

\item{BAT.fun}{The name of the metric or directly it's function.}

\item{BAT.args}{Any named optional arguments to be passed to \code{BAT.metric} (default is \code{NULL})}

\item{return.all}{Whether to return the raw BAT results (\code{TRUE}) or not (\code{FALSE}; default)}
}
\description{
An interface to use metrics from the \code{BAT} package in \code{dispRity}
}
\details{
This function is an interface between the \code{dispRity} and B\code{BAT}AT package allowing to use a \code{BAT} function in the \code{dispRity} pipeline.
By default the function uses the \code{dispRity.metric} format:

\code{metric_value <- BAT.metric(matrix, BAT.fun = BAT::alpha)}

or

\code{metric_value <- BAT.metric(matrix, BAT.fun = "alpha")} 

or

\code{metric_value <- BAT.metric(matrix, BAT.fun = alpha)} 

With any optional argument being handled normally through \code{BAT.args}.

However, most commonly, the function can be used in the context of the \code{dispRity} pipeline as follows:

\code{my_disparity <- dispRity(my_data, metric = BAT.metric, BAT.fun = BAT::alpha)}

\emph{NOTE} that if the \code{dispRity} object contains a \code{$tree} element. It is recycled to the BAT metric if available as an optional argument.
You can override this behaviour by providing your own tree (e.g. using \code{BAT.args = list(tree = my_tree)}) or by removing the tree from your input data (e.g. using \code{remove.tree(my_data)}).
}
\examples{
## Base example:
## Generating a matrix
dummy_matrix <- matrix(rnorm(90), 9, 10)

## Applying a BAT metric to it
alpha_diversity <- BAT.metric(dummy_matrix, BAT.fun = "alpha")

## dispRity example:
## Load ecological data
data(demo_data)
eco_data <- demo_data$jones
## Subseted data based on two groups
eco_data
## Apply the alpha diversity on these subsets
alpha_diversity <- dispRity(eco_data, metric = BAT.metric, BAT.fun = "alpha")
summary(alpha_diversity)

}
\seealso{
\code{\link{dispRity}}, \code{\link{custom.subsets}}
}
\author{
Thomas Guillerme
}
