% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity-package.R
\name{BeckLee}
\alias{BeckLee}
\alias{BeckLee_tree}
\alias{BeckLee_mat50}
\alias{BeckLee_mat99}
\alias{BeckLee_ages}
\title{Beck and Lee 2014 datasets}
\format{
three matrices and one phylogenetic tree.
}
\source{
\url{https://www.royalsocietypublishing.org/doi/abs/10.1098/rspb.2014.1278}
}
\description{
Example datasets from Beck and Lee 2014.
}
\details{
\itemize{
  \item \code{BeckLee_tree} A phylogenetic tree with 50 living and fossil taxa
  \item \code{BeckLee_mat50} The ordinated matrix based on the 50 taxa cladistic distances
  \item \code{BeckLee_mat99} The ordinated matrix based on the 50 taxa + 49 nodes cladistic distances
  \item \code{BeckLee_ages} A list of first and last occurrence data for fossil taxa
  \item \code{BeckLee_disparity} a \code{dispRity} object with estimated sum of variances in 120 time bins, boostrapped 100 times from the Beck and Lee data
}
}
\references{
Beck RMD & Lee MSY. 2014. Ancient dates or accelerated rates?
Morphological clocks and the antiquity of placental mammals.
Proc. R. Soc. B 2014 281 20141278; DOI: 10.1098/rspb.2014.1278
}
\seealso{
BeckLee_disparity disparity
}
