% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.diff.R
\name{char.diff}
\alias{char.diff}
\title{Character differences}
\usage{
char.diff(matrix, method = "hamming")
}
\arguments{
\item{matrix}{A discrete matrix or a list containing discrete characters. The differences is calculated between the columns (usually characters). Use \code{t(matrix)} to calculate the differences between the rows.}

\item{method}{The method to measure difference: \code{"hamming"} (default), \code{"gower"}, \code{"euclidean"}, \code{"ged"} or \code{"mord"}.}
}
\value{
A character difference value or a matrix of class \code{char.diff}
}
\description{
Calculates the character difference from a discrete matrix
}
\details{
The different distances to calculate are:
\itemize{
     \item \code{"hamming"} The Hamming distance: the relative distance between each pairs of comparable characters (i.e. does not interpret the character token significance; the differences are non-ordered Fitch-like: 0-2 = 1).
     \item \code{"gower"} The Gower distance: the absolute distance between each pairs of comparable characters (e.g. does interpret the character token significance; the differences are ordered and absolute: 0-2 = 2)
     \item \code{"euclidean"} The Euclidean distance: @@
     \item \code{"ged"} The Generalised Euclidean Distance (Wills @@): @@
     \item \code{"mord"} The maximum observable distance (Lloyd @@): @@
}

The character difference is calculated as half the sum of the Gower distances between the characters.
}
\examples{
## Comparing two characters
char.diff(list(c(0, 1, 0, 1), c(0, 1, 1, 1)))

## Pairwise comparisons in a morphological matrix
morpho_matrix <- matrix(sample(c(0,1), 100, replace = TRUE), 10)
char.diff(morpho_matrix)

}
\references{
Wills @@
Hamming @@
Lloyd @@
Gower, J.C. 1966. Some distance properties of latent root and vector methods used in multivariate analysis. Biometrika 53:325-338.
}
\seealso{
\code{\link{plot.char.diff}}.
}
\author{
Thomas Guillerme
}
