% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono.test.R
\name{chrono.test}
\alias{chrono.test}
\title{Disparity time series changes}
\usage{
chrono.test(data, method, changepoint, time.window, ...)
}
\arguments{
\item{data}{a \code{"dispRity"} object with time series (e.g. from \code{chrono.subsets}) and disparity values}

\item{method}{the method for measuring change (either \code{"itsa"}, \code{"citsa"}, \code{"area"}, \code{"average"} or any generic \code{h.test} function). See details.}

\item{changepoint}{one or more \code{"numeric"} or \code{"integer"} values of where to measure the changes in disparity time series. This can be set to \code{"detect"} to test all the time subsets available.}

\item{time.window}{optional, either two \code{"numeric"} or \code{"integer"} values to narrow than the time window to perform the test on or a single value that can be either a proportion of subsets to use before and after the \code{changepoint} (up to \code{0.5}) or a number of subsets to use before and after the \code{changepoint}.}

\item{...}{A named list of specific options for each different method. See details.}
}
\description{
Check for changes in disparity time series
}
\details{
The implemented disparity time series changes are:
\itemize{
     \item \code{"itsa"}  @@...
     \item \code{"citsa"} @@... 
     \item \code{"area"}  @@...
     \item \code{"average"}  @@...
}

Each method can take the following specific arguments

\itemize{
     \item \code{"itsa"}  @@...
     \item \code{"citsa"} @@... 
     \item \code{"area"}  @@...
     \item \code{"average"}  @@...
}
}
\seealso{
\code{\link{chrono.subsets}} \code{\link{dispRity}}
}
\author{
Caleb Scutt, Thomas Guillerme
}
