% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Claddis.ordination.R
\name{Claddis.ordination}
\alias{Claddis.ordination}
\title{Imports data from Claddis}
\usage{
Claddis.ordination(data, distance = "MORD", ..., k, add = TRUE, arg.cmdscale)
}
\arguments{
\item{data}{Data from \code{\link[Claddis]{ReadMorphNexus}} or the path to a file to be read by \code{\link[ape]{read.nexus.data}} (see details).}

\item{distance}{Distance type to be computed by \code{\link[Claddis]{MorphDistMatrix}}. Can be either \code{"GC"}, \code{"GED"}, \code{"RED"}, \code{"MORD"}. \code{distance} can also be set to \code{NULL} to convert a matrix in \code{\link[Claddis]{ReadMorphNexus}} list type (see details).}

\item{...}{Any optional arguments to be passed to \code{\link[Claddis]{MorphDistMatrix}}.}

\item{k}{The number of dimensions in the ordination. If left empty, the number of dimensions is set to number of rows - 1.}

\item{add}{whether to use the Cailliez correction for negative eigen values (\code{add = TRUE}; default - see \code{\link[stats]{cmdscale}}) or not (\code{add = FALSE}).}

\item{arg.cmdscale}{Any optional arguments to be passed to \code{\link[stats]{cmdscale}} (as a named list such as \code{list(x.ret = TRUE)}).}
}
\description{
Takes Claddis data and computes both the distance and the ordination matrix
}
\details{
If \code{data} is a file path, the function will use a modified version of \code{\link[ape]{read.nexus.data}} (that handles polymorphic and ambiguous characters). The file content will then be converted into a \code{\link[Claddis]{ReadMorphNexus}} type list treating all characters as unordered.
If the \code{distance} is set to \code{NULL}, \code{data} will be only converted into a \code{\link[Claddis]{ReadMorphNexus}} type list.
}
\examples{
\dontrun{
require(Claddis)

## Ordinating the distance matrix of Claddis example data
Claddis.ordination(Claddis::Michaux1989)

## Creating simple discrete morphological matrix (with polymorphisms)
cat(
"#NEXUS
BEGIN DATA;
DIMENSIONS  NTAX=5 NCHAR=5;
FORMAT SYMBOLS= \" 0 1 2\" MISSING=? GAP=- ;
MATRIX
     t1  {01}1010
     t2  02120
     t3  1210(01)
     t4  01111
     t5  00101
;
END;", file = "morpho_matrix.nex")

## Ordinating the matrix (using a distance matrix)
Claddis.ordination("morpho_matrix.nex")

## Only converting the nexus matrix into a Claddis format
Claddis_data <- Claddis.ordination("morpho_matrix.nex", distance = NULL)

file.remove("morpho_matrix.nex")
}

}
\seealso{
\code{\link[Claddis]{MorphDistMatrix}}, \code{\link[Claddis]{ReadMorphNexus}}, \code{\link[Claddis]{MakeMorphMatrix}}, \code{\link[stats]{cmdscale}}, \code{\link{custom.subsets}}, \code{\link{chrono.subsets}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
