\name{cust.series}
\alias{cust.series}

\title{Separating ordinated data in custom series.}

\description{Splits the ordinated data into a customized series list.}

\usage{
time.series(data, series)
}

\arguments{
\item{data}{can also be a list from time.series!}
\item{series}{some description}
}

\details{If needed}

\seealso{\link{time.series} for spiting the ordinated data in time series.}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}