% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.subsamples.R
\name{custom.subsamples}
\alias{cust.series}
\alias{cust.subsamples}
\alias{custom.series}
\alias{custom.subsamples}
\title{Separating ordinated data in custom subsamples.}
\usage{
custom.subsamples(data, group)
}
\arguments{
\item{data}{An ordinated matrix of maximal dimensions \eqn{k*(k-1)}.}

\item{group}{Either a \code{list} of row numbers or names to be used as different groups or a \code{data.frame} with the same \eqn{k} elements as in \code{data} as rownames.}
}
\value{
This function outputs a \code{dispRity} object containing:
\item{data}{A \code{list} of the split ordinated data (each element is a \code{matrix}).}
\item{elements}{A \code{vector} containing all the rownames from the input matrix.}
\item{subsamples}{A \code{vector} containing the name of the subsamples.}
\code{dispRity} objects can be summarised using \code{print} (S3).
}
\description{
Splits the ordinated data into a customized subsamples list.
}
\details{
Note that every elements from the input data can be assigned to multiple groups!
}
\examples{
## Generating a dummy ordinated matrix
ordinated_matrix <- matrix(data = rnorm(90), nrow = 10)

## Splitting the ordinated matrix into two groups using row numbers
custom.subsamples(ordinated_matrix, list(c(1:4), c(5:10)))

## Splitting the ordinated matrix into three groups using row names
ordinated_matrix <- matrix(data = rnorm(90), nrow = 10,
     dimnames = list(letters[1:10]))
custom.subsamples(ordinated_matrix,
     list("A" = c("a", "b", "c", "d"), "B" = c("e", "f", "g", "h", "i", "j"),
          "C" = c("a", "c", "d", "f", "h")))

## Splitting the ordinated matrix into four groups using a data frame
groups <- as.data.frame(matrix(data = c(rep(1,5), rep(2,5), rep(c(1,2), 5)),
     nrow = 10, ncol = 2, dimnames = list(letters[1:10], c("g1", "g2"))))
custom.subsamples(ordinated_matrix, groups)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{time.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}

