% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispRity-package.R
\docType{package}
\name{dispRity-package}
\alias{dispRity-package}
\title{A package for measuring disparity in R}
\description{
A modulable package for measuring disparity from ordinated matrices.
}
\author{
Thomas Guillerme <guillert@tcd.ie>
}
\keyword{disparity,ordination,phylogeny,cladistic,morphometric,ecology}

