% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.BAT.R
\name{dispRity.BAT}
\alias{dispRity.BAT}
\title{Converts \code{dispRity} to \code{BAT}.}
\usage{
dispRity.BAT(data, subsets, matrix, tree)
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{dispRity} object.}

\item{subsets}{Optional, some specific subsets to extract (see \code{\link{get.subsets}}).}

\item{matrix}{Optional, some specific matrices to extract (see \code{\link{get.matrix}}).}

\item{tree}{Optional, some specific trees to extract (see \code{\link{get.tree}}).}
}
\value{
\itemize{
     \item \code{comm}: a \code{matrix} of presence/absence (\code{0, 1}) sorting the subsets by rows and the elements by columns that can be passed as the \code{comm} argument to \code{BAT} functions. If the \code{data} contains no subsets, the matrix is matrix full of 1 with one row and a number of columns corresponding to the number of elements in \code{data}.
     \item \code{traits}: a \code{matrix} that is the traitspace with elements as rows and dimensions as columns.
     \item \code{tree}: either \code{NULL} if the \code{data} contains no tree or a \code{phylo} or \code{multiPhylo} object from \code{data}.
}
}
\description{
Converts a \code{dispRity} object into \code{BAT} package arguments.
}
\details{
Converts the content of a \code{dispRity} object into a list of arguments that can be used by \code{BAT} functions.
}
\examples{
## Base example:
}
