% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispRity.metric.R
\name{dispRity.metric}
\alias{centroids}
\alias{dispRity.metric}
\alias{mode.val}
\alias{ranges}
\alias{variances}
\title{Disparity metrics}
\usage{
class.metric(matrix)
summary.metric(X)
}
\arguments{
\item{matrix}{An ordinated matrix.}

\item{X}{A vector of \code{class.metric} results.}
}
\description{
Different implemented disparity metrics.
}
\details{
These are inbuilt functions for calculating disparity. The currently implemented class.metric are:
\itemize{
  \item \code{ranges}: calculates the range of each axis of the matrix.
  \item \code{variances}: calculates the variance of each axis of the matrix.
  \item \code{centroids}: calculates the euclidean distance between each row and the centroid of the matrix.
}
The currently implemented summary.metric are:
\itemize{
  \item \code{mode.val}: calculates the modal value of a vector.
}
See also \code{\link[base]{mean}}, \code{\link[stats]{median}}, \code{\link[base]{sum}} or \code{\link[base]{prod}} for commonly used summary metrics.
}
\examples{
##
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{make.metric}}.
}

