% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.metric.R
\name{dispRity.metric}
\alias{centroids}
\alias{centroids.global}
\alias{centroids.observed}
\alias{convhull.surface}
\alias{convhull.volume}
\alias{diagonal}
\alias{dispRity.metric}
\alias{ellipse.volume}
\alias{hyper.volume}
\alias{metric.modifier}
\alias{mode.val}
\alias{ranges}
\alias{variances}
\title{Disparity metrics}
\usage{
level3.fun(matrix, ...)
level2.fun(matrix, ...)
level1.fun(X, ...)
}
\arguments{
\item{matrix}{A matrix.}

\item{X}{A vector.}

\item{...}{Optional arguments to be passed to the function.}
}
\description{
Different implemented disparity metrics.
}
\details{
These are inbuilt functions for calculating disparity. See \code{\link{make.metric}} for details on \code{level3.fun}, \code{level2.fun} and \code{level1.fun}.

The currently implemented \code{level3} metrics are:
\itemize{
  \item \code{ellipse.volume}: calculates the ellipsoid volume of a matrix.
     \itemize{
         \item WARNING: this function only calculates the exact volume from MDS or PCO (PCoA) ordinations (e.g. \code{\link[stats]{cmdscale}}, \code{\link[ape]{pcoa}})
     }
  \item \code{convhull.surface}: calculates the convex hull hyper surface of a matrix.
  \item \code{convhull.volume}: calculates the convex hull hyper volume of a matrix.
     \itemize{
         \item Both \code{convhull} functions are based on the \code{\link[geometry]{convhulln}} function
         \item WARNING: both \code{convhull} functions can be computationally intensive!
     }
  \item \code{hyper.volume}: calculates the hyper volume using the \code{\link[hypervolume]{hypervolume}} algorithm. If no optional argument is given, the different arguments are set by default to:
     \itemize{
         \item \code{repsperpoint = 1000}
         \item \code{bandwidth = \link[hypervolume]{estimate_bandwidth}(matrix, method = "silverman")}
         \item \code{quantile = 0.95}
         \item \code{verbose = FALSE}
         \item \code{warnings = FALSE}
         \item \code{name = NULL}
     }
  \item \code{diagonal}: calculates the longest distance in the ordinated space.
     \itemize{
         \item WARNING: This function is the generalisation of Pythagoras Theorem and thus \bold{works only if each dimensions are orthogonal to each other}.
     }
}
The currently implemented vector aggregate metrics (\code{level2.fun}) are:
\itemize{
  \item \code{ranges}: calculates the range of each axis of the matrix.
  \item \code{variances}: calculates the variance of each axis of the matrix.
  \item \code{centroids}: calculates the euclidean distance between each row and the centroid of the matrix. This function can take an optional arguments \code{centroid} for defining the centroid (if missing (default), the centroid of the matrix is used).
}
The currently implemented value aggregate metrics (\code{level1.fun}) are:
\itemize{
  \item \code{mode.val}: calculates the modal value of a vector.
}
See also \code{\link[base]{mean}}, \code{\link[stats]{median}}, \code{\link[base]{sum}} or \code{\link[base]{prod}} for commonly used summary metrics.
}
\examples{
## A dummy matrix
dummy_matrix <- matrix(rnorm(25), 5, 5)

## variances of a each column in the matrix
variances(dummy_matrix)

## ranges of each column in a matrix
ranges(dummy_matrix)

## Distances between each row and centroid of the matrix
centroids(dummy_matrix)
## Distance between each rows and an arbitrary centroid
centroids(dummy_matrix, centroid = c(0,0,0,0,0))

## Modal value of a vector
mode.val(rnorm(25))

## Ellipsoid volume of a matrix
ellipse.volume(dummy_matrix) # WARNING: only valid for MDS/PCO matrices

## Convex hull hyper-surface of a matrix
convhull.surface(dummy_matrix)

## Convex hull volume of a matrix
convhull.volume(dummy_matrix)

## Matrix hypervolume
hyper.volume(dummy_matrix)

## Matrix diagonal
diagonal(dummy_matrix) # WARNING: only valid if the dimensions are orthogonal

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{make.metric}}.
}

