% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispRity.R
\name{dispRity}
\alias{dispRity}
\title{Calculates disparity from an ordinated matrix.}
\usage{
dispRity(data, metric, verbose = FALSE)
}
\arguments{
\item{data}{An ordinated matrix of maximal dimensions \eqn{k*(k-1)}, a list of matrices (typically output from the functions \code{\link{time.series}} or \code{\link{cust.series}}) or a boostrapped matrix output from \code{\link{boot.matrix}}.}

\item{metric}{A vector containing the class metric and the summary metric (see details).}

\item{verbose}{A \code{logical} value indicating whether to be verbose or not.}
}
\value{
This function outputs a \code{dispRity} object containing:
\item{data}{A \code{list} of the observed and boostraped matrices.}
\item{disparity}{A \code{list} of disparity values.}
\item{elements}{A \code{vector} containing all the names of the elements from the original matrix.}
\item{series}{A \code{vector} containing the name of the series (is \code{"1"} if the input was a single \code{matrix}).}
\item{call}{A \code{vector} containing the arguments used for the bootstraping.}
\code{dispRity} objects can be summarised using \code{print} (S3).
Use \link{summary.dispRity} to summarise the \code{dispRity} object.
}
\description{
Calculates disparity on an ordinated matrix or series of matrices, where the disparity metric can be user specified.
}
\details{
\code{metric} should be input as a vector containing both the class and the summary metric.
The class metric is the descriptor of the matrix (e.g. ranges of the dimensions or distance between taxa and centroid).
The summary metris is the metric used to summarise the matrix descriptor (e.g. the sum of the ranges of the dimensions or the median distance between taxa and centroid).
For example use \code{metric = c(sum, ranges)} for using the sum of the ranges of the dimensions as a disparity metric.
Details on the implemented metric (class and summary) can be found at \code{\link{dispRity.metric}}.
Note that metrics can be also user specified.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity as the sum of ranges from a single matrix
sum_of_ranges <- dispRity(BeckLee_mat50, metric = c(sum, ranges))
summary(sum_of_ranges)
## Bootstraping this value
bootstrapped_data <- boot.matrix(BeckLee_mat50, bootstraps=100)
dispRity(bootstrapped_data, metric=c(sum, ranges))

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2), rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1, dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps=100)
## Caculating the sum of ranges
sum_of_ranges <- dispRity(bootstrapped_data, metric=c(sum, ranges))
summary(sum_of_ranges)
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{boot.matrix}} for bootstraping the data; \code{\link{dispRity.metric}} for details on the implemented metrics and \code{\link{summary.dispRity}} for summarising \code{dispRity} objects.
}

