% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.dispRity.R
\name{extract.dispRity}
\alias{extract.dispRity}
\title{Extracting disparity.}
\usage{
extract.dispRity(data, observed = TRUE, rarefaction)
}
\arguments{
\item{data}{A \code{dispRity} object containing disparity results.}

\item{observed}{A \code{logical} value indicating whether to output the observed (\code{TRUE}) or the bootstrapped values (\code{FALSE}).}

\item{rarefaction}{Either a rarefaction value or "max" or "min" to extract the rarefaction levels (is ignored if \code{observed = TRUE}).}
}
\description{
Extracts the disparity from a disparity object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat99) ; data(BeckLee_tree)

## Calculating some disparity
series <- time.series(BeckLee_mat99, BeckLee_tree, method = "continuous",
     time = c(100, 80, 60), model = "gradual")
bootstraps_dat <- boot.matrix(series, bootstraps = 20, rarefaction = TRUE)
disparity_data <- dispRity(bootstraps_dat, metric = mean)

## Extracting the observed disparity
extract.dispRity(disparity_data)

## Extracting the bootstrapped disparity
boot_disp <- extract.dispRity(disparity_data, observed = FALSE)
str(boot_disp)
## Or with the minimum rarefaction
boot_disp_rare <- extract.dispRity(disparity_data, observed = FALSE,
     rarefaction = "min")
str(boot_disp_rare)
## Or a certain level of rarefaction
boot_disp_rare <- extract.dispRity(disparity_data, observed = FALSE,
     rarefaction = 5)
str(boot_disp_rare)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}; \code{\link{get.dispRity}}.
}

