% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.ordination.R
\name{geomorph.ordination}
\alias{geomorph.ordination}
\title{Imports data from geomorph}
\usage{
geomorph.ordination(data, ...)
}
\arguments{
\item{data}{An array (p x k x n) typically obtained from a procrustes superimposition \code{\link[geomorph]{gpagen}} or a \code{geomorph.data.frame} object.}

\item{...}{Any optional arguments to be passed to \code{\link[stats]{prcomp}}.}
}
\description{
Takes geomorph procrustes object or a geomorph.data.frame object and performs an ordination.
}
\details{
If \code{data} is a \code{geomorph.data.frame} object containing factors, directly performs a \code{\link{custom.subsamples}} using these factors
}
\seealso{
\code{\link[geomorph]{gpagen}}, \code{\link[stats]{prcomp}}, \code{\link{custom.subsamples}}, \code{\link{time.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
