% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.utilities.R
\name{get.contrast.matrix}
\alias{get.contrast.matrix}
\title{Generates a contrast matrix.}
\usage{
get.contrast.matrix(matrix)
}
\arguments{
\item{matrix}{a discrete morphological character matrix.}
}
\description{
Creates a contrast matrix using the observed character states in an input matrix.
}
\examples{
## Create a random matrix
tree <- rcoal(10)
## setting up the parameters
## A gamma rate distribution with of shape alpha = 0.5
my_rates = c(rgamma, 2, 1)
## A fixed substitution rate of 2 (T/T ratio in HKY)
my_substitutions = c(runif, 2, 2)
## Mk matrix (10*50) (for Mkv models)
matrixMk <- sim.morpho(tree, characters = 50, model = "ER", rates = my_rates) 

## Get the matrix's contrast matrix
get.contrast.matrix(matrixMk)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{check.morpho}}
}

