% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{get.series.dispRity}
\alias{get.dispRity}
\alias{get.series.dispRity}
\title{Extracts series from a dispRity object.}
\usage{
get.series.dispRity(data, series)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{series}{A list of series names or series numbers to be extracted.}
}
\value{
This function outputs a \code{dispRity} object.
}
\description{
Extracting some series and data from a \code{dispRity} object.
}
\examples{
## Load the disparity data based on Beck & Lee 2014
data(disparity)

## Get one series
get.series.dispRity(disparity, "60")

## Get two series
get.series.dispRity(disparity, c(1,5))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{extract.dispRity}}.
}

