% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.test.R
\name{lda.test}
\alias{lda.test}
\title{Run a discriminant analysis type}
\usage{
\method{lda}{test}(data, train, prior, type = "linear", bootstraps,
  CV = FALSE, LASSO = FALSE, PGLS = FALSE, all.levels = TRUE, ...)
}
\arguments{
\item{data}{A \code{dispRity} object with attributed subsets or a \code{matrix} with factors as the last column.}

\item{train}{The size of the training dataset. This can be either a single value (to apply to all factors) or a vector of values equal to the number of factors.}

\item{prior}{A \code{vector} of prior expectations (if missing the prior is set to the observed proportions in the training set).}

\item{type}{The type of discriminant function to use: either \code{"linear"} (default), \code{"quadratic"} or any function of similar to the generic type \code{\link[MASS]{lda}}.}

\item{bootstraps}{Optional, the number of bootstraps replicates to run (is missing, no bootstrapss are run).}

\item{CV}{Logical, whether to perform cross-validation or not (default is \code{FALSE}).}

\item{LASSO}{Logical, whether to perform a LASSO analysis.}

\item{PGLS}{Logical, whether to perform a PGLS analysis.}

\item{all.levels}{Logical, whether to always include all levels in the training dataset (\code{TRUE} - default) or not (\code{FALSE}). Not including speeds up the calculation but can generate errors when apply the discriminant function.}

\item{...}{Optional arguments to be passed to the discriminant function.}
}
\value{
A list of class \code{"dispRity"} and \code{"lda.test"} containing the results for each tested factor(s) and each bootstraps along with a \code{$support} element containing meta-data.

For example, for two tested factors with three bootstrap replicates each:

\code{lda_out$factor1$bootstraps1}
\code{               $bootstraps2}
\code{               $bootstraps3}
\code{       $factor2$bootstraps1}
\code{               $bootstraps2}
\code{               $bootstraps3}
\code{       $support}
}
\description{
Running a discriminant analysis or canonical variable analysis test on a \code{dispRity} object
}
\examples{
## The classic iris example:

## A data.frame with the factor as the last column
iris_data <- data.frame(rbind(iris3[,,1], iris3[,,2], iris3[,,3]),
                        species = rep(c("s","c","v"), rep(50,3)))

## The lda test with a training sample size of 50 specimen
iris_test <- lda.test(iris_data, train = 50)

## Printing the results
iris_test

## Summarising and plotting the results
summary(iris_test)
plot(iris_test)

\dontrun{
## A geomorph example
set.seed(1)
require(geomorph)

## Loading the plethodon dataset
data(plethodon)

## Running a Procrustes superimposition
procrustes <- geomorph::gpagen(plethodon$land)

## Creating two factors: the morpho groups and a random group
morpho_group <- as.factor(c(rep("group1", 10), rep("group2", 10), rep("group3", 20)))
random_group <- as.factor(sample(c("random1", "random2", "random3"), 40, replace = TRUE))

## Combining them into a geomorph.data.frame object
geomorph_df <- geomorph.data.frame(procrustes, random = random_group, morpho = morpho_group)

## Converting it into a dispRity object without ordination
disparity_data <- geomorph.ordination(geomorph_df, ordinate = FALSE)

## Applying a simple lda test with a training sample size of 10 specimen bootstrapped 100 times
plethodon_test <- lda.test(disparity_data, train = 10, bootstraps = 100)

## Printing the results
plethodon_test

## Summarising and plotting the results
summary(plethodon_test)
plot(plethodon_test)
}

}
\seealso{
\code{\link{dispRity}}, \code{\link{test.dispRity}}, \code{\link{summary.dispRity}}, \code{\link{plot.dispRity}}.
}
\author{
Thomas Guillerme
}
