% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{make.dispRity}
\alias{make.dispRity}
\title{Creates a \code{dispRity} object.}
\usage{
make.dispRity(data, call, subsamples)
}
\arguments{
\item{data}{A \code{matrix}.}

\item{call}{Optional, a \code{list} to be a \code{dispRity} call.}

\item{subsamples}{Optional, a \code{list} to be a \code{dispRity} subsamples list.}
}
\description{
Creating an empty \code{dispRity} object from a matrix
}
\examples{
## An empty dispRity object
make.dispRity()

## Still an empty dispRity object (with a matrix)
make.dispRity(data = matrix(rnorm(12), ncol = 3))


}
\author{
Thomas Guillerme
}

