% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.metric.R
\name{make.metric}
\alias{make.metric}
\title{Creating disparity metrics}
\usage{
make.metric(fun, ..., silent = FALSE)
}
\arguments{
\item{fun}{Your very own \code{function}!}

\item{...}{Some arguments to be passed to \code{fun}.}

\item{silent}{\code{logical}; if \code{FALSE} (default), the function will be verbose and give no output; if \code{TRUE}, the function will only output the function level.}
}
\description{
Testing the level of disparity metrics
}
\details{
This function tests:
\itemize{
  \item 1: if your function can deal with a matrix as an \code{input}.
  \item 2: which level is your function (level 1, 2 or 3, see \code{\link{dispRity.metric}}).
  \item 3: whether the function can properly be implemented in the \code{dispRity} function.
}
The 3 different metric levels correspond to the dimensions of the output and are:
\itemize{
  \item "level 1": for functions that decompose a \code{matrix} into a single value.
  \item "level 2": for functions that decompose a \code{matrix} into a \code{vector}.
  \item "level 3": for functions that transforme the \code{matrix} into anothrer \code{matrix}.
}
For example, the disparity metric \code{\link[base]{sum}} of \code{\link[dispRity]{variances}} is composed of two metric levels:
\itemize{
  \item The \code{\link[dispRity]{variances}} (level 2) that calculates the variances per column in a matrix (aggregates a \code{matrix} into a \code{vector}).
  \item The \code{\link[base]{sum}} (level 1) that transform the \code{vector} of variances into a single value.
}
See function example for a concrete illustration (three different level of the function \code{\link[base]{sum}}).
}
\examples{
## A level 1 function
my_fun <- function(x) sum(x)
make.metric(my_fun)

## A level 2 function
my_fun <- function(x) apply(x, 2, sum)
make.metric(my_fun)

## A level 3 function
my_fun <- function(x) (x + sum(x))
make.metric(my_fun)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{dispRity.metric}}.
}

