% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.metric.R
\name{make.metric}
\alias{make.metric}
\title{Creating disparity metrics}
\usage{
make.metric(fun)
}
\description{
IN DEVELOPEMENT
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{dispRity.metric}}.
}

