% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.metric.R
\name{make.metric}
\alias{make.metric}
\title{Creating disparity metrics}
\usage{
make.metric(fun, ..., silent = FALSE, check.serial = FALSE, data.dim)
}
\arguments{
\item{fun}{A \code{function}.}

\item{...}{Some arguments to be passed to \code{fun}.}

\item{silent}{\code{logical}; if \code{FALSE} (default), the function will be verbose and give no output; if \code{TRUE}, the function will only output the function's dimension-level.}

\item{check.serial}{\code{logical}; if \code{TRUE}, the function will output a named list containing the metric level and a logical indicating whether the metric can be used serially or not. If \code{FALSE} (default) the function only outputs the metric level.}

\item{data.dim}{optional, two \code{numeric} values for the dimensions of the matrix to run the test function testing. If missing, a default 5 rows by 4 columns matrix is used.}
}
\description{
Testing the dimension-level of disparity metrics
}
\details{

}
\examples{
## A dimension-level 1 function
my_fun <- function(matrix) sum(matrix)
make.metric(my_fun)

## A dimension-level 2 function
my_fun <- function(matrix) apply(matrix, 2, sum)
make.metric(my_fun)

## A dimension-level 3 function
my_fun <- function(matrix) (matrix + sum(matrix))
make.metric(my_fun)

}
\seealso{
\code{\link{dispRity}}, \code{\link{dispRity.metric}}.
}
\author{
Thomas Guillerme
}
