% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.R
\name{model.test}
\alias{model.test}
\title{Disparity Model Testing}
\usage{
model.test(data, models, parameters)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{models}{one or more model names, can be \code{"EB"}, \code{"BM"}, \code{"OU"}, \code{"Trend"} and/or \code{"Shift"}.}

\item{parameters}{a \code{list} of model parameters to be passed to each model.}
}
\description{
Test one or more models of disparity evolution
}
\details{
\itemize{
     \item{\code{"EB"}} an Early-Burst model (see \link{\code{model.test.EB}} for more details)
     \item{\code{"BM"}} a Brownian-Motion model (see \link{\code{model.test.BM}} for more details)
     \item{\code{"OU"}} an Ornstein-Uhlenbeck model (see \link{\code{model.test.OU}} for more details)
     \item{\code{"Trend"}} a Trend model (see \link{\code{model.test.Trend}} for more details)
     \item{\code{"Shift"}} a model with a shift between any of the above models (see \link{\code{model.test.Shift}} for more details)
}
}
\examples{

}
\seealso{
\link{\code{test.dispRity}}, \link{\code{model.test.EB}}, \link{\code{model.test.BM}}, \link{\code{model.test.OU}}, \link{\code{model.test.Trend}}, \link{\code{model.test.Shift}}
}
\author{
Mark Puttick and Thomas Guillerme
}
