% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.R
\name{model.test}
\alias{model.test}
\title{Model Test}
\usage{
model.test(input_disparity, model, pool.variance = NULL,
  time.split.model = NULL, return.model.full = FALSE,
  plot.disparity = TRUE, fixed.optima = FALSE, control.list = list(fnscale
  = -1))
}
\arguments{
\item{input_disparity}{A dispRity object used to test models of evolution through time}

\item{model}{The named model of evolution to allow for changes in disparity-through-time using a homogenous or hetergenous model. A single model can be used as input or multiple modes can be supplied in a vector. If a vector with multiple modes is supplied then the model will test for shifts in modes at the time supplied by 'time.split'. If a list is supplied different models will be fit to the data (See \bold{Details})}

\item{pool.variance}{If NULL the difference in variances will be calculated using Bartlett's Test of equal variances. If there is no significant difference among variances, then variance in samples will be pooled and the same variance will be used for all samples. A significance difference will not pool variances and the original variance will be used for model-testing. If argument TRUE or FALSE are used, Bartlett's test will be ignored and the analyses will use the user-set pooling of variances}

\item{return.model.full}{Logical argument to return full parameter values from each tested model (default = FALSE)}

\item{plot.disparity}{Logical argument to plot the measure of disparity through time and a barplot of the relative support for each model (default = FALSE)}

\item{fixed.optima}{Logical to use an estimated optimum value in OU models (fixed.optima = FALSE), or whether to set the OU optimum to the ancestral value (fixed.optima = TRUE)}

\item{control.list}{Fine-tune control inputs for the optim function}

\item{time.split}{The age of the change in mode. The age is measured as the time before the most recent sample, and multiple ages can be supplied in a vector. If no age is supplied for models then all possible time shifts are fit in the model, and the highest likelihood model is returned. Note this only applies to heterogenous models (See \bold{Details})}
}
\description{
Fit models of disparity change through time
}
\examples{
## To Add
}
\references{
To Add: Hunt 2006, Hunt 2008, Harmon 2010,
}
\seealso{
\code{\link{model.test.sim}}.
}
\author{
Mark N Puttick and Thomas Guillerme
}
