% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.test.sim.R
\name{model.test.sim}
\alias{model.test.sim}
\title{Simulate Model Test}
\usage{
model.test.sim(sim = 1, model, time.split = NULL, time.span = 100,
  variance = 1, sample.size = 100, parameters = list(),
  fixed.optima = FALSE, model.rank = 1, alternative = "two-sided")
}
\arguments{
\item{sim}{The number of separate simulations.}

\item{model}{Either (i) the named model of evolution to simulate for changes in disparity-through-time using a homogenous or hetergenous model (see list in \code{\link{model.test}}) or (ii) an object of class \code{dispRity} returned from \code{model.test} function. If a \code{dispRity} object is supplied, all remaining arguments apart from \code{sim} and \code{model.rank} are ignored as the model specified by the input model is used}

\item{time.split}{The age of the change in mode. The age is measured as the time before the most recent sample, and multiple ages can be supplied in a vector. Note this only applies to heterogenous models}

\item{time.span}{The length of the sequence (\code{numeric}). If one number is supplied this is treated as the length of the sequence and the time span is treated as sequence from 0 to \code{time.span} in unit increments. If the a vector of length > 1 is supplied, this is treated as the the age of each sample in the sequence.}

\item{variance}{The variance of each sample (\code{numeric}). If one number is supplied this is the variance for all samples in the sequence. If a vector of equal length to the \code{time.span} vector is supplied, this is used for the variance of each sample in the sequence}

\item{sample.size}{The sample size of each sample (\code{numeric}). If one number is supplied this is the sample size for all samples in the sequence. If a vector of equal length to the \code{time.span} vector is supplied, this is used for the sample size of each sample in the sequence}

\item{parameters}{A \code{list} of model parameters used for simulations. See details.}

\item{fixed.optima}{A \code{logical} value, whether to use an estimated optimum value in OU models (\code{FALSE} - default), or whether to set the OU optimum to the ancestral value (\code{TRUE}).}

\item{model.rank}{If a \code{dispRity} object is supplied, which model is used for simulation. The rank refers to the order of models as specified by AICc, so if \code{model.rank = 1} (default) the best-fitting model is used for simulation.}

\item{alternative}{If the simulation is based on a \code{dispRity} object, what is the alternative hypothesis. Can be \code{"two-sided"} (default), \code{"greater"} or \code{"lesser"}.}
}
\value{
A list of class \code{dispRity}. Each list element contains the simulated central tendency, as well as the variance, sample size, and subsets used to simulate the data.
}
\description{
Simulate models of disparity change through time
}
\details{
The \code{parameters} is a list of arguments to be passed to the models.
These arguments can be:
\itemize{
     \item{\code{ancestral.state}}, ancestral value of the disparity applicable to all models (default = \code{0.01}).
     \item{\code{sigma.squared}}, rate of step variance to all models except Stasis (default = \code{1}).
     \item{\code{alpha}}, strength of attraction to the optimum in OU models (default = \code{1}).
     \item{\code{optima.1}}, the value of the optimum in a OU model, or the first bin optimum in a multi-OU model (default = \code{0.15}).
     \item{\code{optima.2}}, the second bin optimum in a multi-OU model (default = \code{0.15}).
     \item{\code{optima.3}}, the third bin optimum in a multi-OU model (default = \code{0.15}).
     \item{\code{theta.1}}, the mean in a Stasis model, or the first bin mean in a multi-Stasis model (default = \code{1}).
     \item{\code{theta.2}}, the second bin optimum in a multi-OU model (default = \code{1}).
     \item{\code{theta.3}}, the third bin optimum in a multi-OU model (default = \code{1}).
     \item{\code{omega}}, the variance in a Stasis model (default = \code{1}).
     \item{\code{trend}}, the trend parameter in the Trend model (default = \code{0.5}).
     \item{\code{eb.rate}}, the rate of exponential rate decrease in the EB model (default = \code{-0.1}).
}
}
\examples{
 
 ## Mammal disparity through time
 data(BeckLee_disparity)
 models <- list("Trend", "BM")
 model.test.output <- model.test(BeckLee_disparity, models, time.split = 66)
 
 ## simulations using the output from model.test
 model.test.sim.output <- model.test.sim(sim=10000, model=model.test.output)
  
 plot(model.test.sim.output)
 plot(BeckLee_disparity, add=T, col=c("pink", "#ff000050", "#ff000050"))

 ## simulations using input model parameters
  model.sim.output <- model.test.sim(sim=1000, model="BM", time.span=120, variance=0.1, sample.size=100, parameters=list(ancestral.state=0, sigma.squared=0.1))
 # plot(model.sim.output)
 # plot.dispRity(BeckLee_disparity, add=T, col=c("pink", "#ff000050", "#ff000050"))

}
\references{
Blomberg SP, Garland T Jr, & Ives AR. 2003. Testing for phylogenetic signal in comparative data: behavioral traits are more labile. Evolution.  \bold{57}, 717-745.

Hansen TF. 1997. Stabilizing selection and the comparative analysis of adaptation. Evolution. \bold{51}, 1341-1351.

Harmon LJ, \emph{et al}. 2010. Early bursts of body size and shape evolution are rare in comparative data. \bold{64}, 2385-2396.

Hunt G. 2006. Fitting and comparing models of phyletic evolution: random walks and beyond. Paleobiology. \bold{32}, 578-601. DOI: 10.1666/05070.1.

Hunt G, Hopkins MJ & Lidgard S. 2015. Simple versus complex models of trait evolution and stasis as a response to environmental change. Proceedings of the National Academy of Sciences. \bold{112}, 4885-4890. DOI: 10.1073/pnas.1403662111

Felsenstein J. 1973. Maximum-likelihood estimation of evolutionary trees from continuous characters. American Journal of Human Genetics. \bold{25}, 471-492.

Felsenstein J. 1985. Phylogenies and the comparative method. The American Naturalist. \bold{51}, 1-15.

Murrell DJ. 2018. A global envelope test to detect non‐random bursts of trait evolution. Methods in Ecology and Evolution. DOI: 10.1111/2041-210X.13006

Citation for the envelope code:
}
\seealso{
\code{\link{model.test}}.
}
\author{
Mark N Puttick and Thomas Guillerme
}
