\name{plot.dispRity}
\alias{plot.disparity}

\title{Plotting a dispRity object.}

\description{Plots the dispRity object.}

\usage{
plot.dispRity(dispRity, CI=c(50,95), diversity=FALSE)
}

\arguments{
\item{dispRity}{a dispRity object}
\item{CI}{the confidence interval(s)}
\item{diversity}{optional, whether to plot taxonomic diversity as well}

\details{If needed}

\seealso{\link{dispRity}}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}