% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.dispRity.R
\name{plot.dispRity}
\alias{plot.dispRity}
\title{dispRity object plotting}
\usage{
\method{plot}{dispRity}(data, type, CI = c(50, 95), cent.tend = mean,
  rarefaction = FALSE, diversity = FALSE, ylim, xlab, ylab, col,
  discrete_type = "box", ...)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{type}{Either \code{"continuous"} or \code{"discrete"}.}

\item{CI}{The confidence intervals values (default is \code{CI = c(50,95)}; is ignored if the \code{dispRity} object is not bootstrapped).}

\item{cent.tend}{A function for summarising the bootstrapped disparity values (default is \code{\link[base]{mean}}).}

\item{rarefaction}{Either a \code{logical} whether to rarefy the data; or an \code{integer} for setting a specific rarefaction level or \code{"plot"} to plot the rarefaction curves.}

\item{diversity}{\code{logica} whether to plot the diversity levels (i.e. the number of rows in the matrix).}

\item{discrete_type}{Either \code{"box"} for boxplots or \code{"line"} for distribution lines.}

\item{...}{Any optional arguments to be passed to \code{\link[graphics]{plot}}.}
}
\description{
Plots a \code{dispRity} object.
}
\details{
\code{type}:
\itemize{
  \item \code{"continuous"}: plots the results in a continuous fashion (e.g. disparity function of time)
  \item \code{"discrete"}:plots the results in a discrete fashion (e.g. disparity function of factors)
}
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_tree) ; data(BeckLee_mat50) ; data(BeckLee_mat99) ; data(BeckLee_ages)

## Setting the data
## Generate 5 equidistant time slices in the data set assuming gradual evolutionary models
sliced_data <- time.series(data = BeckLee_mat99, tree = BeckLee_tree, method = "continuous", model = "acctran", time = c(120,90,60,30,0), FADLAD = BeckLee_ages)
bootstrapped_data <- boot.matrix(sliced_data, bootstraps = 20, rarefaction = TRUE)
sum_of_ranges <- dispRity(bootstrapped_data, metric = c(sum, range))

## Discrete plotting
plot(sum_of_ranges, type = "continuous")
## Summarising the results
summary(sum_of_ranges) # default
## Using different options
summary(sum_of_ranges, CI=75, cent.tend=median, rounding=0)
## Recalling the dispRity parameters
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}
}

