% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dispRity.R
\name{plot.dispRity}
\alias{plot.dispRity}
\title{dispRity object plotting}
\usage{
plot.dispRity(data, type, quantile = c(50, 95), cent.tend = mean,
  rarefaction = FALSE, elements = FALSE, ylim, xlab, ylab, col,
  time.series = TRUE, observed = FALSE, add = FALSE, density = NULL,
  ...)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{type}{Either \code{"continuous"} (\code{"c"}), \code{"box"} (\code{"b"}), \code{"lines"} (\code{"l"}) or \code{"polygon"} (\code{"p"}). When unspecified, is set to \code{"continuous"} if \code{\link{time.series}} is used with \code{method = "continuous"}, else is set to \code{"box"}. See details.}

\item{quantile}{The quantiles to display (default is \code{quantile = c(50, 95)}; is ignored if the \code{dispRity} object is not bootstrapped).}

\item{cent.tend}{A function for summarising the bootstrapped disparity values (default is \code{\link[base]{mean}}).}

\item{rarefaction}{Either a \code{logical} whether to rarefy the data; or an \code{integer} for setting a specific rarefaction level or \code{"plot"} to plot the rarefaction curves.}

\item{elements}{\code{logical} whether to plot the number of elements per series.}

\item{time.series}{\code{logical} whether to handle continuous data from the \code{time.series} function as time (in Ma).}

\item{observed}{\code{logical} whether to plot the observed values or not (if existing; default is \code{FALSE}).}

\item{add}{\code{logical} whether to add the new plot an existing one (default is \code{FALSE}).}

\item{density}{the density of shading lines to be passed to \code{link[graphics]{polygon}}. Is ignored if \code{type = "box"} or \code{type = "lines"}.}

\item{...}{Any optional arguments to be passed to \code{\link[graphics]{plot}}.}
}
\description{
Plots a \code{dispRity} object.
}
\details{
\code{type}:
\itemize{
  \item \code{"continuous"}: plots the results as a continuous line.
  \item \code{"box"}: plots the results as discrete box plots (note that this option ignores the user set quantiles and central tendency).
  \item \code{"lines"}: plots the results as discrete vertical lines with the user's set quantiles and central tendency.
  \item \code{"polygon"}: identical as \code{"lines"} but using polygons rather than vertical lines.
}
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_tree) ; data(BeckLee_mat50)
data(BeckLee_mat99) ; data(BeckLee_ages)

## Setting the data
## Generate 5 equidistant time slices in the data set assuming gradual
## evolutionary models
sliced_data <- time.series(data = BeckLee_mat99, tree = BeckLee_tree,
     method = "continuous", model = "acctran", time = 5,
     FADLAD = BeckLee_ages)
bootstrapped_data <- boot.matrix(sliced_data, bootstraps = 20,
     rarefaction = TRUE)
sum_of_variances <- dispRity(bootstrapped_data, metric = c(sum, variances))

## Discrete plotting
plot(sum_of_variances, type = "box")

## Using polygons rather than boxes (quantiles and central tendency can be
## set by the user)
plot(sum_of_variances, type = "polygon", quantile = c(10, 50, 95),
     cent.tend = mode.val)

## Using different options
plot(sum_of_variances, type = "lines", elements = TRUE, ylim = c(0, 5),
     xlab = ("Time (Ma)"), ylab = "disparity", ylim = c(0, 5))

## Continuous plotting (all default options)
plot(sum_of_variances, type = "continuous")

## Using different options (with non time.slicing option)
plot(sum_of_variances, type = "continuous", time.series = FALSE,
     elements = TRUE, col = c("red", "orange", "yellow"))

## Rarefactions plots
plot(sum_of_variances, rarefaction = "plot")

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{summary.dispRity}}.
}

