% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.dispRity.R
\name{print.dispRity}
\alias{print.dispRity}
\title{Prints a \code{dispRity} object.}
\usage{
\method{print}{dispRity}(x, all = FALSE, ...)
}
\arguments{
\item{x}{A \code{dispRity} object.}

\item{all}{\code{logical}; whether to display the entire object (\code{TRUE}) or just summarise it's content (\code{FALSE} - default).}
}
\description{
Summarises the content of a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 matrix
data(BeckLee_mat50)

## Creating a dispRity object
disparity_object <- disparity(BeckLee_mat50, metric = c(sum, ranges))

## Displaying the summary of the object content
disparity_object
print(disparity_object)
print.dispRity(disparity_object)

## Displaying the full object
print.dispRity(disparity_object, all = TRUE)
}
\author{
Thomas Guillerme
}

