% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.model.test.R
\name{rank.model.test}
\alias{rank.model.test}
\title{rank.model.test}
\usage{
rank.model.test(sim = 1000, model)
}
\arguments{
\item{sim}{The number of separate simulations.}

\item{model}{An object of class \code{dispRity} returned from \code{model.test} function.}
}
\value{
A matrix with the relative fit, parameter values, and Rank Envelope test p values for each model
}
\description{
A wrapper function for \code{\link{model.test.sim}} to summarise relative fit of multiple models from \code{\link{model.test}}. Gives the relative fit of \code{model.test} output using log-likelihood and AICc values, as well as a significance test to elucidate if empirical data is significantly different to simulated data modelled using the estimated model parameters from \code{model.test}
}
\examples{

## Mammal disparity through time
data(BeckLee_disparity)
models <- list("BM", "OU", "multi.OU", "Trend")
models.out <- model.test(BeckLee_disparity, models, time.split = 66)

## assess the significance of simulated data against the empirical distribution
rank.model.test(sim=1000, model=models.out)

}
\references{
Murrell DJ. 2018. A global envelope test to detect non-random bursts of trait evolution. Methods in Ecology and Evolution. DOI: 10.1111/2041-210X.13006
}
\seealso{
\code{\link{model.test}} and \code{\link{model.test.sim}}.
}
\author{
Mark N Puttick and Thomas Guillerme
}
