% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.utilities.R
\name{scale.dispRity}
\alias{scale.dispRity}
\title{Scaling and centering disparity results.}
\usage{
\method{scale}{dispRity}(data, center = FALSE, scale = FALSE,
  use.all = TRUE)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{center}{either a \code{logical} value or a \code{numeric} vector of length equal to the number of elements of \code{data} (default is \code{FALSE}).}

\item{scale}{either a \code{logical} value or a \code{numeric} vector of length equal to the number of elements of \code{data} (default is \code{FALSE}).}

\item{use.all}{\code{logical}, whether to scale/center using the full distribution (i.e. all the disparity values) or only the distribution within each series of bootstraps (default is \code{TRUE}).}
}
\description{
Scales or/and centers the disparity measurements.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Setting up two customised series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps = 7,
     rarefaction = c(10, 25))
## Calculating the sum of centroids
disparity <- dispRity(bootstrapped_data, metric = c(sum, centroids))

## Scaling the data
summary(disparity) # No scaling
summary(scale(disparity)) # Dividing by the maximum
## Multiplying by 10 (dividing by 0.1)
summary(scale.dispRity(disparity, max = 0.1))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{test.dispRity}}, \code{link[base]{scale}}.
}

