% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.dispRity.R
\name{scale.dispRity}
\alias{scale.dispRity}
\title{Scaling disparity results.}
\usage{
\method{scale}{dispRity}(data, max)
}
\arguments{
\item{data}{a \code{dispRity} object.}

\item{max}{an optional value for scaling. If missing the maximum disparity from \code{data} is used.}
}
\description{
Divides calculated disparity by the maximum disparity.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculate disparity as the sum of the centroids
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps = 7,
     rarefaction = c(10, 25))
## Calculating the sum of variances
disparity <- dispRity(bootstrapped_data, metric = c(sum, centroids))

## Scaling the data
summary(disparity)
summary(scale.dispRity(disparity))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}} and \code{\link{test.dispRity}}.
}

