% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential.test.R
\name{sequential.test}
\alias{sequential.test}
\title{Sequential linear regressions}
\usage{
sequential.test()
}
\description{
Function still in development.
}
\seealso{
\code{\link{test.dispRity}}, \code{\link{bhatt.coeff}}, \code{\link{null.test}}.
}
\author{
Thomas Guillerme
}
