% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.tree.R
\name{slice.tree}
\alias{slice.tree}
\title{Slicing a tree.}
\usage{
slice.tree(tree, age, model, FAD, LAD)
}
\arguments{
\item{tree}{A \code{phylo} object with a \code{root.time} element.}

\item{age}{A single \code{numeric} value where to perform the slice.}

\item{model}{One of the following models: \code{"acctran"}, \code{"deltran"}, \code{"punctuated"} or \code{"gradual"}. Is ignored if \code{method = "discrete"}.}

\item{FAD, LAD}{The first and last occurrence data.}
}
\description{
Slicing through a phylogenetic tree (function modified from paleotree::timeSliceTree).
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link[paleotree]{timeSliceTree}} and \code{\link{time.series}}.
}

