% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.tree.R
\name{slice.tree}
\alias{slice.tree}
\title{Slicing a tree.}
\usage{
slice.tree(tree, age, model, FAD, LAD)
}
\arguments{
\item{tree}{A \code{phylo} object with a \code{root.time} element.}

\item{age}{A single \code{numeric} value where to perform the slice.}

\item{model}{One of the following models: \code{"acctran"}, \code{"deltran"}, \code{"punctuated"} or \code{"gradual"}. Is ignored if \code{method = "discrete"}.}

\item{FAD, LAD}{The first and last occurrence data.}
}
\description{
Slicing through a phylogenetic tree (function modified from paleotree::timeSliceTree).
}
\examples{
set.seed(1)
## Generate a random ultrametric tree
tree <- rcoal(20)

## Add some node labels
tree$node.label <- letters[1:19]

## Add it's root time
tree$root.time <- max(tree.age(tree)$ages)

## Create a slice on the tree at age 0.5
tree_75 <- slice.tree(tree, age = 0.75, "deltran")

}
\seealso{
\code{\link[paleotree]{timeSliceTree}}, \code{\link{time.subsamples}}.
}
\author{
Thomas Guillerme
}
