% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.dispRity.R
\name{sort.dispRity}
\alias{sort.dispRity}
\title{Sorts a \code{dispRity} object.}
\usage{
\method{sort}{dispRity}(data, decreasing = FALSE, sort)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{decreasing}{\code{logical}. Should the sort be increasing or decreasing? Not available for partial sorting}

\item{sort}{an optional \code{vector} of \code{numeric} values corresponding to the order in which to return the series..}
}
\description{
Sorts the series of a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat99) ; data(BeckLee_tree) 

## Split the data
series <- time.series(data = BeckLee_mat99, tree = BeckLee_tree,
      method = "continuous", time = 5, model = "acctran")

## Calculating the sum of centroids
disparity <- dispRity(series, metric = c(sum, centroids))

## Sorting the data
summary(disparity)
#summary(sort.dispRity(disparity, decreasing = TRUE))
#summary(sort(disparity, sort = c(1,3,4,5,2)))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{test.dispRity}}, \code{\link{plot.dispRity}}, \code{\link{get.dispRity}}, \code{\link{extract.dispRity}}.
}

