% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dispRity.R
\name{summary.dispRity}
\alias{summary.dispRity}
\title{dispRity object summary}
\usage{
\method{summary}{dispRity}(data, quantiles = c(50, 95), cent.tend = mean,
  recall = FALSE, rounding, results = "coefficients")
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{quantiles}{The quantiless to display (default is \code{quantiles = c(50,95)}; is ignored if the \code{dispRity} object is not bootstrapped).}

\item{cent.tend}{A function for summarising the bootstrapped disparity values (default is \code{\link[base]{mean}}).}

\item{recall}{\code{logical}, whether to recall the \code{dispRity} parameters input (default = \code{FALSE}).}

\item{rounding}{Optional, a value for rounding the values in the output table (default = 2).}

\item{results}{Optional, in the case of summarising a \code{\link{sequential.test}} which results to display (default = "coefficients")}
}
\value{
A \code{data.frame} with:
\item{series}{the series names.}
\item{n}{the number of elements per series.}
\item{observed}{the observed disparity or the the observed central tendency (<cent_tend>) of disparity (\code{obs.<cent_tend>}).}
\item{bootstraps...}{if \code{data} is bootstrapped, the bootstrapped disparity's central tendency (\code{bs.<cent_tend>}) and the quantiless of the bootstrapped disparity's (or, if \code{data} is not bootstrapped but disparity is calculated as a distribution - see \code{\link[dispRity]{dispRity}}) - the quantiless of the observed disparity is displayed).}
}
\description{
Creates a summary of a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps = 100)
## Calculating the sum of variances
sum_of_sum_of_variances <- dispRity(bootstrapped_data,
     metric = c(sum, sum_of_variances))

## Summarising the results
summary(sum_of_sum_of_variances) # default
## Using different options
summary(sum_of_sum_of_variances, quantiles = 75, cent.tend = median,
     rounding = 0, recall = TRUE)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}
}

