\name{summary.dispRity}
\alias{summary.disparity}

\title{Summarizing a dispRity object.}

\description{Summarize the dispRity object.}

\usage{
summary.dispRity(dispRity, CI=c(50,95))
}

\arguments{
\item{dispRity}{a dispRity object}
\item{CI}{the confidence interval(s)}

\details{If needed}

\seealso{\link{dispRity}}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}