% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dispRity.R
\name{summary.dispRity}
\alias{summary.dispRity}
\title{dispRity object summary}
\usage{
\method{summary}{dispRity}(data, quantiles = c(50, 95), cent.tend = median,
  recall = FALSE, rounding)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{quantiles}{The quantiles to display (default is \code{quantiles = c(50, 95)}; is ignored if the \code{dispRity} object is not bootstrapped).}

\item{cent.tend}{A function for summarising the bootstrapped disparity values (default is \code{\link[stats]{median}}).}

\item{recall}{\code{logical} value specifying whether to recall the \code{dispRity} parameters input (default = \code{FALSE}).}

\item{rounding}{Optional, a value for rounding the values in the output table (default = 2).}
}
\value{
A \code{data.frame} with:
\item{subsets}{the subset names.}
\item{n}{the number of elements in each subset.}
\item{observed}{the observed disparity or the the observed central tendency (<cent_tend>) of disparity (\code{obs.<cent_tend>}).}
\item{bootstraps...}{if \code{data} is bootstrapped, the bootstrapped disparity's central tendency (\code{bs.<cent_tend>}) and the quantiles of the bootstrapped disparities (or, if \code{data} is not bootstrapped but disparity is calculated as a distribution - see \code{\link[dispRity]{dispRity}}) - the quantiles of the observed disparity are displayed).}
}
\description{
Creates a summary of a \code{dispRity} object.
}
\examples{
## Load the disparity data based on Beck & Lee 2014
data(disparity)

## Summarising the results
summary(disparity) # default
## Using different options
summary(disparity, quantiles = 75, cent.tend = mean, rounding = 8,
     recall = TRUE)

}
\seealso{
\code{\link{dispRity}}, \code{\link{plot.dispRity}}.
}
\author{
Thomas Guillerme
}
