% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.dispRity.R
\name{summary.dispRity}
\alias{summary.dispRity}
\title{dispRity object summary}
\usage{
\method{summary}{dispRity}(data, CI = c(50, 95), cent.tend = mean,
  recall = FALSE, rounding)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{CI}{The confidence intervals values (default is \code{CI = c(50,95)}; is ignored if the \code{dispRity} object is not bootstrapped).}

\item{cent.tend}{A function for summarising the bootstrapped disparity values (default is \code{\link[base]{mean}}).}

\item{recall}{\code{logical}, whether to recall the \code{dispRity} parameters input.}

\item{rounding}{Optional, a value for rounding the central tendency and the confidence intervals in the output table.}
}
\description{
Creates a summary of a \code{dispRity} object.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2), rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1, dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps=100)
## Caculating the sum of ranges
sum_of_ranges <- dispRity(bootstrapped_data, metric=c(sum, range))

## Summarising the results
summary(sum_of_ranges) # default
## Using different options
summary(sum_of_ranges, CI=75, cent.tend=median, rounding=0, recall=TRUE)
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}
}

