% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.dispRity.R
\name{test.dispRity}
\alias{test.dispRity}
\title{Disparity statistics}
\usage{
test.dispRity(data, test, comparisons, ...)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{test}{A statistical \code{function} to apply to the data.}

\item{comparisons}{If data contains more than two series, the type of comparisons to apply: either \code{"referential"}, \code{"sequential"}, \code{"pairwise"} or an even \code{vector} of series names/number to compare (see details).}

\item{...}{Additional options to pass to the test \code{function}.}

\item{format}{The expected input format to the function (either \code{"matrix"} or \code{"vector"} - default).}
}
\description{
Applying statistical tests to dispRity objects
}
\details{
The \code{comparison} argument can be:
\itemize{
  \item \code{"referential"}: compares the first series to all the others.
  \item \code{"sequential"}: compares each series sequentially (e.g. first against second, second against third, etc.).
  \item \code{"pairwise"}: pairwise comparisons of all the series.
  \item An even \code{vector} of names or number of series to compare.
}
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, nrow(BeckLee_mat50)/2),
     rep(2, nrow(BeckLee_mat50)/2)), nrow = nrow(BeckLee_mat50), ncol = 1,
     dimnames = list(rownames(BeckLee_mat50))))
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps=100)
## Caculating the sum of ranges
sum_of_ranges <- dispRity(bootstrapped_data, metric=c(sum, ranges))

## Summarising the results
summary(sum_of_ranges) # default
## Using different options
summary(sum_of_ranges, quantile = 75, cent.tend = median,
     rounding=  0, recall = TRUE)
}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}
}

