% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.dispRity.R
\name{test.dispRity}
\alias{test.dispRity}
\title{Disparity statistics}
\usage{
test.dispRity(data, test, comparisons = "pairwise", correction, ...,
  details = FALSE)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{test}{A test \code{function} to apply to the data.}

\item{comparisons}{If data contains more than two series, the type of comparisons to apply: either \code{"pairwise"} (default), \code{"referential"}, \code{"sequential"}, \code{"all"} or a list of pairs of series names/number to compare (see details).}

\item{correction}{which p-value correction to apply to \code{htest} category test (see \code{\link[stats]{p.adjust}}). If missing, no correction is applied.}

\item{...}{Additional options to pass to the test \code{function}.}

\item{details}{Whether to output the details of each test (non-formatted; default = \code{FALSE}).}
}
\description{
Applying statistical tests to dispRity objects
}
\details{
The \code{comparison} argument can be:
\itemize{
  \item \code{"pairwise"}: pairwise comparisons of all the series (default).
  \item \code{"referential"}: compares the first series to all the others.
  \item \code{"sequential"}: compares each series sequentially (e.g. first against second, second against third, etc.).
  \item \code{"all"}: compares all the series simultaneously:
     \itemize{
         \item If the test is of type \code{\link[stats]{aov}} or \code{\link[stats]{glm}} the applied formula will be \code{bootstrapped disparity ~ series names}.
         \item See the \code{\link[dispRity]{sequential.test}} for a \code{\link[stats]{glm}} correcting for time autocorrelation.
     }
  \item A list of pairs of number of series to compare. Each element of the the list must contain two elements
     (e.g. \code{list(c("a","b"), ("b", "a"))} to compare "a" to "b" and then "b" to "a").
}
IMPORTANT: if you are performing multiple comparisons (e.g. when using \code{"pairwise"}, \code{"referential"} or \code{"sequential"}),
don't forget about the Type I error rate inflation. You might want to use a \emph{p-value} correction (see \code{\link[stats]{p.adjust}}).
NOTE: some specific test like \code{\link[stats]{aov}} or \code{\link[dispRity]{sequential.test}} ignore the comparisons argument.
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)

## Calculating the disparity from a customised series
## Generating the series
factors <- as.data.frame(matrix(data = c(rep(1, 12), rep(2, 13), rep(3, 25)),
      dimnames =list(rownames(BeckLee_mat50))), ncol = 1)
customised_series <- cust.series(BeckLee_mat50, factors)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_series, bootstraps = 100)
## Calculating the sum of variances
sum_of_variances <- dispRity(bootstrapped_data, metric = c(sum, variances))

## Measuring the series overlap
test.dispRity(sum_of_variances, bhatt.coeff, "pairwise")

## Measuring differences from a reference_series
test.dispRity(sum_of_variances, wilcox.test, "referential")

## Testing the effect of the factors
test.dispRity(sum_of_variances, aov, "all")
## warning: this violates some aov assumptions!

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{dispRity}}, \code{\link{dispRity.test}}
}

