% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.dispRity.R
\name{test.dispRity}
\alias{test.dispRity}
\title{Testing disparity hypotheses}
\usage{
test.dispRity(data, test, comparisons = "pairwise", rarefaction = NULL,
  correction = "none", concatenate = TRUE, conc.quantiles = c(mean, c(95,
  50)), details = FALSE, ...)
}
\arguments{
\item{data}{A \code{dispRity} object.}

\item{test}{A test \code{function} to apply to the data.}

\item{comparisons}{If data contains more than two subsets, the type of comparisons to apply: either \code{"pairwise"} (default), \code{"referential"}, \code{"sequential"}, \code{"all"} or a list of pairs of subset names/number to compare (see details).}

\item{rarefaction}{A \code{numeric} value indicating whether to use a specific rarefaction level (default = \code{NULL}).}

\item{correction}{Which p-value correction to apply to \code{htest} category test (see \code{\link[stats]{p.adjust}}; default = \code{"none"}).}

\item{concatenate}{Logical, whether to concatenate bootstrapped disparity values (\code{TRUE}; default) or to apply the test to each bootstrapped value individually (\code{FALSE}).}

\item{conc.quantiles}{If \code{concatenate = TRUE}, must be a central tendency function and a vector of quantiles (default = \code{c(mean, c(95, 50))}).}

\item{details}{Whether to output the details of each test (non-formatted; default = \code{FALSE}).}

\item{...}{Additional options to pass to the test \code{function}.}
}
\description{
Applying statistical tests to dispRity objects
}
\details{
The \code{comparison} argument can be:
\itemize{
  \item \code{"pairwise"}: pairwise comparisons of all the subsets (default).
  \item \code{"referential"}: compares the first subset to all the others.
  \item \code{"sequential"}: compares each subset sequentially (e.g. first against second, second against third, etc.).
  \item \code{"all"}: compares all the subsets simultaneously to the data (i.e. \code{bootstrapped disparity ~ subsets names}). This argument is used for \code{\link[stats]{lm}} or \code{\link[stats]{glm}} type tests.
  \item A list of pairs of number of subsets to compare. Each element of the list must contain two elements
     (e.g. \code{list(c("a","b"), ("b", "a"))} to compare "a" to "b" and then "b" to "a").
  \item {If the called test is \code{\link[dispRity]{null.test}}, the comparison argument is ignored.} 
}
IMPORTANT: if you are performing multiple comparisons (e.g. when using \code{"pairwise"}, \code{"referential"} or \code{"sequential"}),  don't forget about the Type I error rate inflation. You might want to use a \emph{p-value} correction (see \code{\link[stats]{p.adjust}}).
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_mat50)
data(BeckLee_tree)

## Calculating the disparity from customised subsets
## Generating the subsets
groups <- crown.stem(BeckLee_tree, inc.nodes = FALSE)
customised_subsets <- custom.subsets(BeckLee_mat50, groups)
## Bootstrapping the data
bootstrapped_data <- boot.matrix(customised_subsets, bootstraps = 100)
## Calculating the sum of variances
sum_of_variances <- dispRity(bootstrapped_data, metric = c(sum, variances))

## Measuring the subset overlap
test.dispRity(sum_of_variances, bhatt.coeff, "pairwise")

## Measuring differences from a reference subset
test.dispRity(sum_of_variances, wilcox.test, "referential")

## Running a linear model on the data
test.dispRity(sum_of_variances, lm, "all")

## Measuring disparity as a distribution
disparity_var <- dispRity(bootstrapped_data, metric = variances)
## Differences between the concatenated bootstrapped values of the subsets
test.dispRity(disparity_var, test = t.test, comparisons = "pairwise",
     concatenate = TRUE, correction = "bonferroni")
## Differences between the subsets bootstrapped
test.dispRity(disparity_var, test = t.test, comparisons = "pairwise",
     concatenate = FALSE, correction = "bonferroni",
     conc.quantiles = c(mean, c(95, 5)))

}
\seealso{
\code{\link{dispRity}}, \code{\link{null.test}}, \code{\link{bhatt.coeff}}, \code{\link{pair.plot}}.
}
\author{
Thomas Guillerme
}
