% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.metric.R
\name{test.metric}
\alias{test.metric}
\title{Test disparity metric}
\usage{
test.metric(
  data,
  metric,
  ...,
  shifts,
  shift.options,
  model,
  replicates = 3,
  steps = 10,
  dimensions,
  verbose = FALSE
)
}
\arguments{
\item{data}{A matrix or a \code{dispRity} object (see details).}

\item{metric}{A vector containing one to three functions. At least of must be a dimension-level 1 or 2 function (see details). If \code{data} is a \code{dispRity} object with disparity already calculated, this argument can be left empty (and the one from \code{data} is recycled)}

\item{...}{Optional arguments to be passed to the metric.}

\item{shifts}{The types of shits to test, can be \code{"random"}, \code{"size"}, \code{"density"} and \code{"position"}.}

\item{shift.options}{Optional, a \code{list} of named arguments to be passed to \code{\link{reduce.space}}}

\item{model}{Optional, which model to fit for testing the metric. See details.}

\item{replicates}{A \code{numeric} number of replicates to increase variance. By default \code{replicates = 3}.}

\item{steps}{The number of steps in the space reduction to output between 1\% and 100\%. By default \code{steps = 10}.}

\item{dimensions}{Optional, a \code{numeric} value or proportion of the dimensions to keep.}

\item{verbose}{A \code{logical} value indicating whether to be verbose or not.}
}
\description{
Test whether a metric captures changes trait space size, density and position.
}
\details{
The default \code{model} is a linear model using the following function:
    \code{model = function(data) lm(reduction \~ disparity, data)}
You can provide your own as long as it is a single function with \code{data} as a single argument. The two terms from data should be called \code{reduction} for the variable on the x axis and \code{disparity} for the variable on the y axis. For example:
    \code{model = function(data) nls(disparity \~ a*reduction/(b+reduction), data)}
Note that models (like this example) should be specific to the dataset. Any type of model can be fitted but only the ones with an associated \code{summary} function will be correctly displayed by \code{\link{summary.test.metric}}.
}
\examples{
## Load a disparity data
data(disparity)

## Testing the metric used in the disparity data 
median_centroid_test <- test.metric(disparity, shifts = c("random", "size"))

## Summarising the results
summary(sum_var_test)

## Visualising the results
plot(sum_var_test)
 
}
\references{
Guillerme T, Puttick MN, Marcy AE, Weisbecker V. \bold{2020} Shifting spaces: Which disparity or dissimilarity measurement best summarize occupancy in multidimensional spaces?. Ecol Evol. 2020;00:1-16. (doi:10.1002/ece3.6452)
}
\author{
Thomas Guillerme
}
