% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.series.R
\name{time.series}
\alias{time.series}
\title{Separating ordinated data in time series.}
\usage{
\method{time}{series}(data, tree, method, time, model, inc.nodes, FADLAD,
  verbose = FALSE)
}
\arguments{
\item{tree}{A \code{phylo} object.}

\item{age}{The age of the tree. If missing the age is set to be the tree height.}

\item{order}{Either "past" if the units is express time since the present (e.g. million years ago), or "present" if the unit is expressed in time since the root.}
}
\description{
Splits the ordinated data into a time series list.
}
\examples{
## A dated random phylogeny with a root 50 units of time old.
tree.age(rtree(10), age=50)
## A random tree with the distance since the root.
tree.age(rtree(10), order='present')
}
\author{
Thomas Guillerme
}

