% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.subsamples.R
\name{time.subsamples}
\alias{time.subsamples}
\alias{time.series}
\title{Separating data in time subsamples.}
\usage{
time.subsamples(data, tree, method, time, model, inc.nodes = FALSE, FADLAD, verbose = FALSE, t0 = FALSE)
}
\arguments{
\item{data}{A matrix.}

\item{tree}{A \code{phylo} object matching the data and with a \code{root.time} element. This argument can be left missing if \code{method = "discrete"} and all elements are present in the optional \code{FADLAD} argument.}

\item{method}{The time subsampling method: either \code{"discrete"} (or \code{"d"}) or \code{"continuous"} (or \code{"c"}).}

\item{time}{Either a single \code{integer} for the number of discrete or continuous samples or a \code{vector} containing the age of each sample.}

\item{model}{One of the following models: \code{"acctran"}, \code{"deltran"}, \code{"punctuated"} or \code{"pro"}. Is ignored if \code{method = "discrete"}.}

\item{inc.nodes}{A \code{logical} value indicating whether nodes should be included in the time subsamples. Is ignored if \code{method = "continuous"}.}

\item{FADLAD}{An optional \code{data.frame} containing the first and last occurrence data.}

\item{verbose}{A \code{logical} value indicating whether to be verbose or not. Is ignored if \code{method = "discrete"}.}

\item{t0}{If \code{time} is a number of samples, whether to start the sampling from the \code{tree$root.time} (\code{TRUE}), or from the first sample containing at least three elements (\code{FALSE} - default) or from a fixed time point (if \code{t0} is a single \code{numeric} value).}
}
\value{
This function outputs a \code{dispRity} object containing:
\item{matrix}{the multidimensional space (a \code{matrix}).}
\item{call}{A \code{list} containing the called arguments.}
\item{subsamples}{A \code{list} containing matrices pointing to the elements present in each subsamples.}

Use \link{summary.dispRity} to summarise the \code{dispRity} object.
}
\description{
Splits the data into a time subsamples list.
}
\details{
If \code{method = "continuous"} and when the sampling is done along an edge of the tree, the data selected for the time subsamples is:
\itemize{
  \item \code{"acctran"}: always the value from the ancestral node.
  \item \code{"deltran"}: always the value from the descendant node or tip.
  \item \code{"punctuated"}: randomly selected from the ancestral node or the descendant node or tip with a 50% probability each.
  \item \code{"proximity"}: selects the ancestral node or the descendant with a probability relative to branch length.
}
}
\examples{
## Load the Beck & Lee 2014 data
data(BeckLee_tree) ; data(BeckLee_mat50)
data(BeckLee_mat99) ; data(BeckLee_ages)

## Time binning (discrete method)
## Generate two discrete time bins from 120 to 40 Ma every 40 Ma
time.subsamples(data = BeckLee_mat50, tree = BeckLee_tree, method = "discrete",
     time = c(120, 80, 40), inc.nodes = FALSE, FADLAD = BeckLee_ages)
## Generate the same time bins but including nodes
time.subsamples(data = BeckLee_mat99, tree = BeckLee_tree, method = "discrete",
     time = c(120, 80, 40), inc.nodes = TRUE, FADLAD = BeckLee_ages)

## Time slicing (continuous method)
## Generate five equidistant time slices in the dataset assuming a proximity
## evolutionary model
time.subsamples(data = BeckLee_mat99, tree = BeckLee_tree,
     method = "continuous", model = "acctran", time = 5,
     FADLAD = BeckLee_ages)

}
\seealso{
\code{\link{tree.age}}, \code{\link{slice.tree}}, \code{\link{cust.subsamples}}, \code{\link{boot.matrix}}, \code{\link{dispRity}}.
}
\author{
Thomas Guillerme
}
