\name{tree.age}
\alias{tree.age}

\title{Calculating the age of nodes and tips in a tree.}

\description{Calculates the age of each nodes and tips in a tree give the hight of the tree or some specified age.}

\usage{
tree.age(arg)
}

\arguments{
\item{arg}{some description}

\details{If needed}

\seealso{\link{disparity} or any other functions}

\examples{
## Some example
1+1
}

\author{Thomas Guillerme}